Crate pspsdk

Modules

allocatorsnon-stub-code

Allocators for the PSP system.

ionon-stub-code

Traits, helpers, and type definitions for core I/O functionality.

osnon-stub-code

OS-specific functionality.

panickingnon-stub-code

Implementation of various bits and pieces of the panic! macro and associated runtime pieces.

powernon-stub-code

Power management abstractions.

processnon-stub-code

A module for working with processes.

syncnon-stub-code

Useful synchronization primitives.

sys

The low-level implementations, type definition, and system stubs for the PSP system.

timenon-stub-code

Temporal quantification.

Macros

call_mainnon-stub-code

Call the psp_main function. handling things on ffi boundary.

dbgnon-stub-code

Prints and returns the value of a given expression for quick and dirty debugging.

default_module_startnon-stub-code

Creates the default module_start implementation.

dprintnon-stub-code

Prints to the PSP screen.

dprintlnnon-stub-code

Prints to the PSP screen, with a newline.

eprintnon-stub-code

Prints to the PSP standard error. Equivalent to the print! macro, except that output goes to io::stderr instead of io::stdout. See print! for example usage.

eprintlnnon-stub-code

Prints to the PSP standard error, with a newline.

export

Creates a single library export entry for a PSP module.

exports

Creates one or more library export entry for a PSP module and “.lib.ent”.

modulenon-stub-code

Declare a PSP module.

module_info

Declare a PSP module info.

printnon-stub-code

Prints to the PSP standard output.

printlnnon-stub-code

Prints to the PSP standard output, with a newline.

psp_fw

Evaluates to a boolean a targeted PSP firmware version match the expected version.

psp_fw_select

Selects code at compile-time based on psp_fw predicates.

Functions

enable_home_buttonnon-stub-code

Enable the home button.

init_cwd⚠non-stub-code and non-std

Initialize the current working directory.

module_start_init⚠non-stub-code and non-std

Initialization of stuff required to live on module_start. Returns argc and argv tuple.

process_argc_argv⚠non-stub-code and non-std

Process argc_bytes and argp from module_start and creates a argc and argv to pass to be passed to be used later. when starting the module main thread.

set_custom_cleanupnon-stub-code and non-std

Sets a custom cleanup procedure that will be called on exit and on the set callback by enable_home_button.

set_psp_os_functionsnon-stub-code

Sets the PSP OS functions for the io module when non-stub-code is enabled.

Attribute Macros

psp_fw_cfg

Conditionally includes the form to which it is attached based on a PSP Firmware version targeted.

psp_stub

Creates a PSP stub library from a extern "C" block.