pspsdk::allocators

Struct VariablePoolAlloc

pub struct VariablePoolAlloc { /* private fields */ }

Available on crate feature non-stub-code only.

A memory allocator using the PSP Variable-sized Memory Pool.

Implementations

impl VariablePoolAlloc

pub fn new(pool_size: usize) -> Result<Self>

Creates a variable memory pool allocator with a given size.

The RAM partition used is the default partition on the compilation context. For user application is MemoryPartitionId::MainUser while for kernel it is MemoryPartitionId::MainKernel. To specify or configure the VariablePoolAlloc further, you can use VariablePoolAlloc::builder and build to your liking.

pub const fn builder() -> VariablePoolAllocBuilder

Creates a builder for the VariablePoolAlloc that can be used to configure it’s behavior before creating the allocator.