pspsdk

Function module_start_init

pub unsafe fn module_start_init(
    argc_bytes: usize,
    argp: *mut c_void,
) -> (usize, [*mut u8; 20])

Available on crate feature non-stub-code and non-crate feature std only.

Initialization of stuff required to live on module_start. Returns argc and argv tuple.

For barebones projects, it is the user responsibility to call this function on module_start.

Safety

argp must be valid for argc_bytes. And this function is only to be used with module_start.