pspsdk::io

Struct ArrayLineWriter

pub struct ArrayLineWriter<W, const N: usize>
where
    W: Write + ?Sized,
{ /* private fields */ }

Available on crate feature non-stub-code only.

Wraps a writer and buffers output to it, flushing whenever a newline (0x0a, '\n') is detected.

The ArrayBufWriter struct wraps a writer and buffers its output. But it only does this batched write when it goes out of scope, or when the internal buffer is full. Sometimes, you’d prefer to write each line as it’s completed, rather than the entire buffer at once. Enter ArrayLineWriter. It does exactly that.

Like ArrayBufWriter, a ArrayLineWriter’s buffer will also be flushed when the ArrayLineWriter goes out of scope or when its internal buffer is full.

If there’s still a partial line in the buffer when the ArrayLineWriter is dropped, it will flush those contents.

Examples

We can use ArrayLineWriter to write one line at a time, significantly reducing the number of actual writes to the file.

use std::fs::{self, File};

use io_core::io::{self, ArrayLineWriter, Write};

fn main() -> io::Result<()> {
    let road_not_taken = b"I shall be telling this with a sigh
Somewhere ages and ages hence:
Two roads diverged in a wood, and I -
I took the one less traveled by,
And that has made all the difference.";

    let file = File::create("poem.txt")?;
    let mut file: ArrayLineWriter<_, 512> = ArrayLineWriter::new(file);

    file.write_all(b"I shall be telling this with a sigh")?;

    // No bytes are written until a newline is encountered (or
    // the internal buffer is filled).
    assert_eq!(fs::read_to_string("poem.txt")?, "");
    file.write_all(b"\n")?;
    assert_eq!(fs::read_to_string("poem.txt")?, "I shall be telling this with a sigh\n",);

    // Write the rest of the poem.
    file.write_all(
        b"Somewhere ages and ages hence:
Two roads diverged in a wood, and I -
I took the one less traveled by,
And that has made all the difference.",
    )?;

    // The last line of the poem doesn't end in a newline, so
    // we have to flush or drop the `ArrayLineWriter` to finish
    // writing.
    file.flush()?;

    // Confirm the whole poem was written.
    assert_eq!(fs::read("poem.txt")?, &road_not_taken[..]);
    Ok(())
}

Implementations

impl<W, const N: usize> ArrayLineWriter<W, N>
where W: Write,

pub const fn new(inner: W) -> ArrayLineWriter<W, N>

Creates a new ArrayLineWriter.

Examples

use std::fs::File;

use io_core::io::{self, ArrayLineWriter};

let file = File::create("poem.txt")?;
let file: ArrayLineWriter<_, 512> = ArrayLineWriter::new(file);
Ok(())

pub const fn get_mut(&mut self) -> &mut W

Gets a mutable reference to the underlying writer.

Caution must be taken when calling methods on the mutable reference returned as extra writes could corrupt the output stream.

Examples

use std::fs::File;

use io_core::io::{self, ArrayLineWriter};

let file = File::create("poem.txt")?;
let mut file = ArrayLineWriter::<_, 512>::new(file);

// we can use reference just like file
let reference = file.get_mut();

pub fn into_inner(self) -> Result<W, IntoInnerError<ArrayLineWriter<W, N>>>

Unwraps this ArrayLineWriter, returning the underlying writer.

The internal buffer is written out before returning the writer.

Errors

An [Err] will be returned if an error occurs while flushing the buffer.

Examples

use std::fs::File;

use io_core::io::{self, ArrayLineWriter};

let file = File::create("poem.txt")?;

let writer: ArrayLineWriter<File, 512> = ArrayLineWriter::new(file);

let file: File = writer.into_inner()?;

impl<W, const N: usize> ArrayLineWriter<W, N>
where W: Write + ?Sized,

pub const fn get_ref(&self) -> &W

Gets a reference to the underlying writer.

Examples

use std::fs::File;

use io_core::io::{self, ArrayLineWriter};

let file = File::create("poem.txt")?;
let file = ArrayLineWriter::<_, 512>::new(file);

let reference = file.get_ref();