pspsdk::io

Struct PipeReader

pub struct PipeReader(/* private fields */);

Available on crate feature non-stub-code only.

Read end of an anonymous pipe.

Implementations

impl PipeReader

pub fn try_clone(&self) -> Result<PipeReader, Error>

Creates a new PipeReader instance that shares the same underlying file description.

Examples

use std::{
    fs,
    io::{pipe, Write},
    process::Command,
};
const NUM_SLOT: u8 = 2;
const NUM_PROC: u8 = 5;
const OUTPUT: &str = "work.txt";

let mut jobs = vec![];
let (reader, mut writer) = pipe()?;

// Write NUM_SLOT characters the pipe.
writer.write_all(&[b'|'; NUM_SLOT as usize])?;

// Spawn several processes that read a character from the pipe, do some work, then
// write back to the pipe. When the pipe is empty, the processes block, so only
// NUM_SLOT processes can be working at any given time.
for _ in 0..NUM_PROC {
    jobs.push(
        Command::new("bash")
            .args(["-c", &format!("read -n 1\necho -n 'x' >> '{OUTPUT}'\necho -n '|'",)])
            .stdin(reader.try_clone()?)
            .stdout(writer.try_clone()?)
            .spawn()?,
    );
}

// Wait for all jobs to finish.
for mut job in jobs {
    job.wait()?;
}

// Check our work and clean up.
let xs = fs::read_to_string(OUTPUT)?;
fs::remove_file(OUTPUT)?;
assert_eq!(xs, "x".repeat(NUM_PROC.into()));