pspsdk::io

Struct Stdin

pub struct Stdin { /* private fields */ }

Available on crate feature non-stub-code only.

A handle to the standard input stream of a process.

Each handle is a shared reference to a global buffer of input data to this process. A handle can be lock’d to gain full access to BufRead methods (e.g., .lines()). Reads to this handle are otherwise locked with respect to other reads.

This handle implements the Read trait, but beware that concurrent reads of Stdin must be executed with care.

Created by the io::stdin method.

Examples

use pspsdk::io;

fn main() -> io::Result<()> {
    let mut buffer = String::new();
    let stdin = io::stdin(); // We get `Stdin` here.
    stdin.read_line(&mut buffer)?;
    Ok(())
}

Implementations

impl Stdin

pub fn lock(&self) -> StdinLock<'static>

Locks this handle to the standard input stream, returning a readable guard.

The lock is released when the returned lock goes out of scope.was The returned guard also implements the Read and BufRead traits for accessing the underlying data.

Examples

use pspsdk::io::{self, BufRead};

fn main() -> io::Result<()> {
    let mut buffer = String::new();
    let stdin = io::stdin();
    let mut handle = stdin.lock();

    handle.read_line(&mut buffer)?;
    Ok(())
}

pub fn read_line(&self, buf: &mut String) -> Result<usize>

Locks this handle and reads a line of input, appending it to the specified buffer.

For detailed semantics of this method, see the documentation on BufRead::read_line.

Examples

use pspsdk::io;

let mut input = String::new();
match io::stdin().read_line(&mut input) {
    Ok(n) => {
        println!("{n} bytes read");
        println!("{input}");
    },
    Err(error) => println!("error: {error}"),
}

You can run the example one of two ways:

Pipe some text to it, e.g., printf foo | path/to/executable

Give it text interactively by running the executable directly, in which case it will wait for the Enter key to be pressed before continuing

pub fn lines(self) -> Lines<StdinLock<'static>>

Consumes this handle and returns an iterator over input lines.

For detailed semantics of this method, see the documentation on BufRead::lines.

Examples

use pspsdk::io;

let lines = io::stdin().lines();
for line in lines {
    println!("got a line: {}", line.unwrap());
}