pspsdk::io

Type Alias RawOsError

pub type RawOsError = i32;

Available on crate feature non-stub-code only.

The type of raw OS error codes returned by Error::raw_os_error.

This is an [i32] on all currently supported platforms, but platforms added in the future (such as UEFI) may use a different primitive type like [usize]. Use asor into conversions where applicable to ensure maximum portability.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 4 bytes