pspsdk

Macro exports

exports!() { /* proc-macro */ }

Creates one or more library export entry for a PSP module and “.lib.ent”.

The syntax rules follows the same logic as [export] macro, but with multiple items separated by ;.

Example

pspsdk::module_info!("example_module", 1, 1);

#[unsafe(no_mangle)]
static EXAMPLE: u32 = 0xBADBAD;

#[unsafe(no_mangle)]
extern "C" fn example() -> i32 {
    42
}

#[unsafe(no_mangle)]
extern "C" fn example2() -> i32 {
    44
}

#[unsafe(no_mangle)]
extern "C" fn module_start(argc_bytes: usize, argv: *mut ::core::ffi::c_void) -> u32 {
    0
}

pspsdk::exports! {
    "syslib", 0, 0, 0x8000, [
        fn module_start,
        static module_info.0 : 0xF01D73A7,
    ];
    "libname", 1, 0, 0, [
        fn example,
        fn example2 : 0xDEADDEAD,
        static EXAMPLE,
    ];
}