pspsdk

Macro psp_fw_select

psp_fw_select!() { /* proc-macro */ }

Selects code at compile-time based on psp_fw predicates.

This macro evaluates, at compile-time, a series of psp_fw predicates, selects the first that is true, and emits the code guarded by that predicate. The code guarded by other predicates is not emitted.

An optional trailing _ wildcard can be used to specify a fallback. If none of the predicates are true, a [compile_error] is emitted.

Examples

pspsdk::psp_fw_select! {
    ..150 => {
        pub fn my_func() -> i32 { /* PSP <1.50 specific functionality */ }
    },
    150..370 => {
        pub fn my_func() -> i32 { /* PSP 1.50~370 specific functionality */ }
    },
    _ => {
        pub fn my_func() -> i32 { /* fallback implementation */ }
    },
}

The psp_fw_select! macro can also be used in expression position, with or without braces on the right-hand side:

let _some_string = pspsdk::psp_fw_select! {
    ..420 => "With great power comes great electricity bills",
    _ => { "Behind every successful diet is an unwatched pizza" }
};

It can also be used on type expression right-hand side:

struct StructV1 {
    // fields
}
struct StructV2 {
    // fields
}

type Struct = pspsdk::psp_fw_select! {
    ..150 => StructV1,
    _ => StructV2,
};