pspsdk::os::fd

Struct BorrowedFd

pub struct BorrowedFd<'fd> { /* private fields */ }

Available on crate feature non-stub-code only.

A borrowed file descriptor.

This has a lifetime parameter to tie it to the lifetime of something that owns the file descriptor. For the duration of that lifetime, it is guaranteed that nobody will close the file descriptor.

This uses repr(transparent) and has the representation of a host file descriptor, so it can be used in FFI in places where a file descriptor is passed as an argument, it is not captured or consumed.

This type does not have a [ToOwned][alloc::borrow::ToOwned] implementation. Calling .to_owned() on a variable of this type will call it on &BorrowedFd and use Clone::clone() like ToOwned does for all types implementing Clone. The result will be descriptor borrowed under the same lifetime.

To obtain an OwnedFd, you can use BorrowedFd::try_clone_to_owned instead, but this is not supported on all platforms.

Implementations

impl BorrowedFd<'_>

pub const unsafe fn borrow_raw(fd: RawFd) -> Self

Returns a BorrowedFd holding the given raw file descriptor.

Safety

The resource pointed to by fd must remain open for the duration of the returned BorrowedFd.

Panics

Panics if the raw file descriptor has the value -1.

pub fn try_clone_to_owned(&self) -> Result<OwnedFd>

Creates a new OwnedFd instance that shares the same underlying file description as the existing BorrowedFd instance.

impl BorrowedFd<'_>

pub fn read(&self, buf: &mut [u8]) -> Result<usize>

pub fn read_buf(&self, cursor: BorrowedCursor<'_, u8>) -> Result<()>

pub fn write(&self, buf: &[u8]) -> Result<usize>

pub fn write_all(&self, buf: &[u8]) -> Result<()>

pub fn seek(&self, pos: SeekFrom) -> Result<u64>

pub fn stream_position(&self) -> Result<u64>