pspsdk::os::fd

Struct OwnedFd

pub struct OwnedFd { /* private fields */ }

Available on crate feature non-stub-code only.

An owned file descriptor.

This closes the file descriptor on drop. It is guaranteed that nobody else will close the file descriptor.

This uses repr(transparent) and has the representation of a host file descriptor, so it can be used in FFI in places where a file descriptor is passed as a consumed argument or returned as an owned value.

You can use AsFd::as_fd to obtain a BorrowedFd.

Implementations

impl OwnedFd

pub fn try_clone(&self) -> Result<Self>

Creates a new OwnedFd instance that shares the same underlying file description as the existing OwnedFd instance.