pspsdk::os::fd

Trait IntoRawFd

pub trait IntoRawFd {
    // Required method
    fn into_raw_fd(self) -> u32;
}

Available on crate feature non-stub-code only.

A trait to express the ability to consume an object and acquire ownership of its raw file descriptor.

Required Methods

fn into_raw_fd(self) -> u32

Consumes this object, returning the raw underlying file descriptor.

This function is typically used to transfer ownership of the underlying file descriptor to the caller. When used in this way, callers are then the unique owners of the file descriptor and must close it once it’s no longer needed.

However, transferring ownership is not strictly required. Use a [Into<OwnedFd>::into] implementation for an API which strictly transfers ownership.

Example

use std::fs::File;
#[cfg(any(unix, target_os = "wasi"))]
use std::os::fd::{IntoRawFd, RawFd};

let f = File::open("foo.txt")?;
#[cfg(any(unix, target_os = "wasi"))]
let raw_fd: RawFd = f.into_raw_fd();

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".