pspsdk

Module panic

Available on crate feature non-stub-code only.

Structs

AssertUnwindSafe

A simple wrapper around a type to assert that it is unwind safe.

Location

A struct containing information about the location of a panic.

PanicHookInfo

A struct providing information about a panic.

PanicInfo

A struct providing information about a panic.

PanicMessage

A message that was given to the panic!() macro.

Traits

RefUnwindSafe

A marker trait representing types where a shared reference is considered unwind safe.

UnwindSafe

A marker trait which represents “panic safe” types in Rust.

Functions

always_abort

Makes all future panics abort directly without running the panic hook or unwinding.

catch_unwind

Invokes a closure, capturing the cause of an unwinding panic if one occurs.

resume_unwind

Triggers a panic without invoking the panic hook.

abort_on_unwindExperimental

Invokes a closure, aborting if the closure unwinds.