pspsdk::panic

Struct PanicHookInfo

pub struct PanicHookInfo<'a> { /* private fields */ }

Available on crate feature non-stub-code only.

A struct providing information about a panic.

PanicHookInfo structure is passed to a panic hook set by the set_hook function.

Examples

use pspsdk::{panic, println};

panic::set_hook(Box::new(|panic_info| {
    println!("panic occurred: {panic_info}");
}));

panic!("critical system failure");

Implementations

impl<'a> PanicHookInfo<'a>

pub fn payload(&self) -> &(dyn Any + Send)

Returns the payload associated with the panic.

This will commonly, but not always, be a &'static str or String. If you only care about such payloads, use payload_as_str instead.

A invocation of the panic!() macro in Rust 2021 or later will always result in a panic payload of type &'static str or String.

Only an invocation of [panic_any] (or, in Rust 2018 and earlier, panic!(x) where x is something other than a string) can result in a panic payload other than a &'static str or String.

Examples

use pspsdk::{panic, println};

panic::set_hook(Box::new(|panic_info| {
    if let Some(s) = panic_info.payload().downcast_ref::<&str>() {
        println!("panic occurred: {s:?}");
    } else if let Some(s) = panic_info.payload().downcast_ref::<String>() {
        println!("panic occurred: {s:?}");
    } else {
        println!("panic occurred");
    }
}));

panic!("Normal panic");

pub fn payload_as_str(&self) -> Option<&str>

Returns the payload associated with the panic, if it is a string.

This returns the payload if it is of type &'static str or String.

A invocation of the panic!() macro in Rust 2021 or later will always result in a panic payload where payload_as_str returns Some.

Only an invocation of [panic_any] (or, in Rust 2018 and earlier, panic!(x) where x is something other than a string) can result in a panic payload where payload_as_str returns None.

Example

pspsdk::panic::set_hook(Box::new(|panic_info| {
    if let Some(s) = panic_info.payload_as_str() {
        pspsdk::println!("panic occurred: {s:?}");
    } else {
        pspsdk::println!("panic occurred");
    }
}));

panic!("Normal panic");

pub fn location(&self) -> Option<&'static Location<'static>>

Returns information about the location from which the panic originated, if available.

This method will currently always return [Some], but this may change in future versions.

Examples

use pspsdk::{panic, println};

panic::set_hook(Box::new(|panic_info| {
    if let Some(location) = panic_info.location() {
        println!("panic occurred in file '{}' at line {}", location.file(), location.line(),);
    } else {
        println!("panic occurred but can't get location information...");
    }
}));

panic!("Normal panic");

pub fn can_unwind(&self) -> bool

Returns whether the panic handler is allowed to unwind the stack from the point where the panic occurred.

This is true for most kinds of panics with the exception of panics caused by trying to unwind out of a Drop implementation or a function whose ABI does not support unwinding.

It is safe for a panic handler to unwind even when this function returns false, however this will simply cause the panic handler to be called again.