pspsdk::panic

Struct PanicInfo

pub struct PanicInfo<'a> { /* private fields */ }

Available on crate feature non-stub-code only.

A struct providing information about a panic.

A PanicInfo structure is passed to the panic handler defined by #[panic_handler].

For the type used by the panic hook mechanism in std, see std::panic::PanicHookInfo.

Implementations

impl<'a> PanicInfo<'a>

pub fn message(&self) -> PanicMessage<'_>

The message that was given to the panic! macro.

Example

The type returned by this method implements Display, so it can be passed directly to write!() and similar macros.

#[panic_handler]
fn panic_handler(panic_info: &PanicInfo<'_>) -> ! {
    write!(DEBUG_OUTPUT, "panicked: {}", panic_info.message());
    loop {}
}

pub fn location(&self) -> Option<&'static Location<'static>>

Returns information about the location from which the panic originated, if available.

This method will currently always return [Some], but this may change in future versions.

Examples

use std::panic;

panic::set_hook(Box::new(|panic_info| {
    if let Some(location) = panic_info.location() {
        println!("panic occurred in file '{}' at line {}",
            location.file(),
            location.line(),
        );
    } else {
        println!("panic occurred but can't get location information...");
    }
}));

panic!("Normal panic");

pub fn payload(&self) -> &(dyn Any + Send + 'static)

👎Deprecated since 1.81.0:
this never returns anything useful

Returns the payload associated with the panic.

On this type, core::panic::PanicInfo, this method never returns anything useful. It only exists because of compatibility with std::panic::PanicHookInfo, which used to be the same type.

See std::panic::PanicHookInfo::payload.

pub fn can_unwind(&self) -> bool

🔬This is a nightly-only experimental API. (panic_can_unwind)

Returns whether the panic handler is allowed to unwind the stack from the point where the panic occurred.

This is true for most kinds of panics with the exception of panics caused by trying to unwind out of a Drop implementation or a function whose ABI does not support unwinding.

It is safe for a panic handler to unwind even when this function returns false, however this will simply cause the panic handler to be called again.