pspsdk

Module panicking

Available on crate feature non-stub-code only.

Implementation of various bits and pieces of the panic! macro and associated runtime pieces.

Specifically, this module contains the implementation of:

Panic hooks

Executing a panic up to doing the actual implementation

Shims around “try”

Functions

begin_panic

This is the entry point of panicking for the non-format-string variants of panic!() and assert!(). In particular, this is the only entry point that supports arbitrary payloads, not just format strings.

catch_unwindNon-panic=immediate-abort and non-std

Invoke a closure, capturing the cause of an unwinding panic if one occurs.

panic_handlerNon-std

Entry point of panics from the core crate (panic_impl lang item).

panicking

Determines whether the current thread is unwinding because of panic.

resume_unwind

This is the entry point for resume_unwind. It just forwards the payload to the panic runtime.