pspsdk::sync

Module nonpoison

Available on crate feature non-stub-code only.

Non-poisoning synchronous locks.

The difference from the locks in the poison module is that the locks in this module will not become poisoned when a thread panics while holding a guard.

Re-exports

pub use super::poison::Once;

Structs

Barrier

A barrier enables multiple threads to synchronize the beginning of some computation.

BarrierWaitResult

A BarrierWaitResult is returned by Barrier::wait() when all threads in the Barrier have rendezvoused.

Condvar

A Condition Variable

LazyLock

A value which is initialized on the first access.

MappedMutexGuard

An RAII mutex guard returned by PoisonMutexGuard::map, which can point to a subfield of the protected data. When this structure is dropped (falls out of scope), the lock will be unlocked.

MappedRwLockReadGuard

RAII structure used to release the shared read access of a lock when dropped, which can point to a subfield of the protected data.

MappedRwLockWriteGuard

RAII structure used to release the exclusive write access of a lock when dropped, which can point to a subfield of the protected data.

Mutex

A mutual exclusion primitive useful for protecting shared data

MutexGuard

An RAII implementation of a “scoped lock” of a mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.

OnceLock

A synchronization primitive which can nominally be written to only once.

ReentrantLock

A re-entrant mutual exclusion lock

ReentrantLockGuard

An RAII implementation of a “scoped lock” of a re-entrant lock. When this structure is dropped (falls out of scope), the lock will be unlocked.

RwLock

A reader-writer lock

RwLockReadGuard

RAII structure used to release the shared read access of a lock when dropped.

RwLockWriteGuard

RAII structure used to release the exclusive write access of a lock when dropped.

Enums

TryLockError

An enumeration of possible errors associated with a TryLockResult which can occur while trying to acquire a lock, from the try_lock method on a Mutex or the try_read and try_write methods on an RwLock.

Type Aliases

TryLockResult

A type alias for the result of a nonblocking locking method.