pspsdk::sync::nonpoison

Struct MappedMutexGuard

pub struct MappedMutexGuard<'a, T: ?Sized + 'a, M: RawMutex = SpinMutex> { /* private fields */ }

Available on crate feature non-stub-code only.

An RAII mutex guard returned by PoisonMutexGuard::map, which can point to a subfield of the protected data. When this structure is dropped (falls out of scope), the lock will be unlocked.

The main difference between MappedMutexGuard and MutexGuard is that the former cannot be used with Condvar, since that could introduce soundness issues if the locked object is modified by another thread while the Mutex is unlocked.

The data protected by the mutex can be accessed through this guard via its [Deref] and [DerefMut] implementations.

This structure is created by the map and try_map methods on MutexGuard.

Implementations

impl<'a, T: ?Sized, M: RawMutex> MappedMutexGuard<'a, T, M>

pub fn map<U, F>(orig: Self, f: F) -> MappedMutexGuard<'a, U, M>
where F: FnOnce(&mut T) -> &mut U, U: ?Sized,

Makes a MappedMutexGuard for a component of the borrowed data, e.g. an enum variant.

The Mutex is already locked, so this cannot fail.

This is an associated function that needs to be used as MappedMutexGuard::map(...). A method would interfere with methods of the same name on the contents of the MutexGuard used through Deref.

pub fn try_map<U, F>( orig: Self, f: F, ) -> Result<MappedMutexGuard<'a, U, M>, Self>
where F: FnOnce(&mut T) -> Option<&mut U>, U: ?Sized,

Makes a MappedMutexGuard for a component of the borrowed data. The original guard is returned as an Err(...) if the closure returns None.

The Mutex is already locked, so this cannot fail.

This is an associated function that needs to be used as MappedMutexGuard::try_map(...). A method would interfere with methods of the same name on the contents of the MutexGuard used through Deref.