pspsdk::sync::nonpoison

Struct ReentrantLock

pub struct ReentrantLock<T: ?Sized, M = SpinMutex> { /* private fields */ }

Available on crate feature non-stub-code only.

A re-entrant mutual exclusion lock

This lock will block other threads waiting for the lock to become available. The thread which has already locked the mutex can lock it multiple times without blocking, preventing a common source of deadlocks.

Examples

Allow recursively calling a function needing synchronization from within a callback (this is how StdoutLock is currently implemented):

use core::cell::RefCell;

use pspsdk::sync::nonpoison::ReentrantLock;

pub struct Log {
    data: RefCell<String>,
}

impl Log {
    pub fn append(&self, msg: &str) {
        self.data.borrow_mut().push_str(msg);
    }
}

static LOG: ReentrantLock<Log> = ReentrantLock::new(Log {
    data: RefCell::new(String::new()),
});

pub fn with_log<R>(f: impl FnOnce(&Log) -> R) -> R {
    let log = LOG.lock();
    f(&*log)
}

with_log(|log| {
    log.append("Hello");
    with_log(|log| log.append(" there!"));
});

Implementations

impl<T> ReentrantLock<T>

pub const fn new(t: T) -> ReentrantLock<T>

Creates a new re-entrant lock in an unlocked state ready for use.

Examples

use pspsdk::sync::nonpoison::ReentrantLock;

let lock = ReentrantLock::new(0);

impl<T, M: RawMutex> ReentrantLock<T, M>

pub const fn new_with(value: T, raw_mutex: M) -> Self

Creates a new re-entrant lock in an unlocked state ready for use using the specified mutex.

pub fn into_inner(self) -> T

Consumes this lock, returning the underlying data.

Examples

use pspsdk::sync::nonpoison::ReentrantLock;

let lock = ReentrantLock::new(0);
assert_eq!(lock.into_inner(), 0);

impl<T: ?Sized, M: RawMutex> ReentrantLock<T, M>

pub fn lock(&self) -> ReentrantLockGuard<'_, T, M>

Acquires the lock, blocking the current thread until it is able to do so.

This function will block the caller until it is available to acquire the lock. Upon returning, the thread is the only thread with the lock held. When the thread calling this method already holds the lock, the call succeeds without blocking.

Examples

use std::{cell::Cell, sync::Arc, thread};

use pspsdk::sync::nonpoison::ReentrantLock;

let lock = Arc::new(ReentrantLock::new(Cell::new(0)));
let c_lock = Arc::clone(&lock);

thread::spawn(move || {
    c_lock.lock().set(10);
})
.join()
.expect("thread::spawn failed");
assert_eq!(lock.lock().get(), 10);

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data.

Since this call borrows the ReentrantLock mutably, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist.

Examples

use pspsdk::sync::nonpoison::ReentrantLock;

let mut lock = ReentrantLock::new(0);
*lock.get_mut() = 10;
assert_eq!(*lock.lock(), 10);

pub const fn data_ptr(&self) -> *const T

Returns a raw pointer to the underlying data.

The returned pointer is always non-null and properly aligned, but it is the user’s responsibility to ensure that any reads and writes through it are properly synchronized to avoid data races, and that it is not read or written through after the mutex is dropped.