pspsdk::sync::poison

Enum TryLockError

pub enum TryLockError<T> {
    Poisoned(PoisonError<T>),
    WouldBlock,
}

Available on crate feature non-stub-code only.

An enumeration of possible errors associated with a TryLockResult which can occur while trying to acquire a lock, from the try_lock method on a Mutex or the try_read and try_write methods on an RwLock.

Variants

Poisoned(PoisonError<T>)

The lock could not be acquired because another thread failed while holding the lock.

WouldBlock

The lock could not be acquired at this time because the operation would otherwise block.