pspsdk::sync::poison

Struct Flag

pub struct Flag { /* private fields */ }

Available on crate feature non-stub-code only.

Implementations

impl Flag

pub const fn new() -> Flag

pub fn borrow(&self) -> LockResult<()>

Check the flag for an unguarded borrow, where we only care about existing poison.

pub fn guard(&self) -> LockResult<Guard>

Check the flag for a guarded borrow, where we may also set poison when done.

pub fn done(&self, guard: &Guard)

pub fn get(&self) -> bool

pub fn clear(&self)