pspsdk::sync::poison

Struct Mutex

pub struct Mutex<T: ?Sized, M = SpinMutex> { /* private fields */ }

Available on crate feature non-stub-code only.

A mutual exclusion primitive useful for protecting shared data

This mutex will block threads waiting for the lock to become available. The mutex can be created via a new constructor. Each mutex has a type parameter which represents the data that it is protecting. The data can only be accessed through the RAII guards returned from lock and try_lock, which guarantees that the data is only ever accessed when the mutex is locked.

Poisoning

The mutexes in this module implement a strategy called “poisoning” where a mutex is considered poisoned whenever a thread panics while holding the mutex. Once a mutex is poisoned, all other threads are unable to access the data by default as it is likely tainted (some invariant is not being upheld).

For a mutex, this means that the lock and try_lock methods return a [Result] which indicates whether a mutex has been poisoned or not. Most usage of a mutex will simply unwrap() these results, propagating panics among threads to ensure that a possibly invalid invariant is not witnessed.

A poisoned mutex, however, does not prevent all access to the underlying data. The PoisonError type has an into_inner method which will return the guard that would have otherwise been returned on a successful lock. This allows access to the data, despite the lock being poisoned.

Examples

use std::{
    sync::{mpsc::channel, Arc},
    thread,
};

use pspsdk::sync::poison::Mutex;

const N: usize = 10;

// Spawn a few threads to increment a shared variable (non-atomically), and
// let the main thread know once all increments are done.
//
// Here we're using an Arc to share memory among threads, and the data inside
// the Arc is protected with a mutex.
let data = Arc::new(Mutex::new(0));

let (tx, rx) = channel();
for _ in 0..N {
    let (data, tx) = (Arc::clone(&data), tx.clone());
    thread::spawn(move || {
        // The shared state can only be accessed once the lock is held.
        // Our non-atomic increment is safe because we're the only thread
        // which can access the shared state when the lock is held.
        //
        // We unwrap() the return value to assert that we are not expecting
        // threads to ever fail while holding the lock.
        let mut data = data.lock();
        *data += 1;
        if *data == N {
            tx.send(()).unwrap();
        }
        // the lock is unlocked here when `data` goes out of scope.
    });
}

rx.recv().unwrap();

To recover from a poisoned mutex:

use std::{sync::Arc, thread};

use pspsdk::sync::poison::Mutex;

let lock = Arc::new(Mutex::new(0_u32));
let lock2 = Arc::clone(&lock);

let _ = thread::spawn(move || -> () {
    // This thread will acquire the mutex first, unwrapping the result of
    // `lock` because the lock has not been poisoned.
    let _guard = lock2.lock();

    // This panic while holding the lock (`_guard` is in scope) will poison
    // the mutex.
    panic!();
})
.join();

// The lock is poisoned by this point, but the returned result can be
// pattern matched on to return the underlying guard on both branches.
let mut guard = match lock.lock_checked() {
    Ok(guard) => guard,
    Err(poisoned) => poisoned.into_inner(),
};

*guard += 1;

To unlock a mutex guard sooner than the end of the enclosing scope, either create an inner scope or drop the guard manually.

use std::{sync::Arc, thread};

use pspsdk::sync::poison::Mutex;

const N: usize = 3;

let data_mutex = Arc::new(Mutex::new(vec![1, 2, 3, 4]));
let res_mutex = Arc::new(Mutex::new(0));

let mut threads = Vec::with_capacity(N);
(0..N).for_each(|_| {
    let data_mutex_clone = Arc::clone(&data_mutex);
    let res_mutex_clone = Arc::clone(&res_mutex);

    threads.push(thread::spawn(move || {
        // Here we use a block to limit the lifetime of the lock guard.
        let result = {
            let mut data = data_mutex_clone.lock();
            // This is the result of some important and long-ish work.
            let result = data.iter().fold(0, |acc, x| acc + x * 2);
            data.push(result);
            result
            // The mutex guard gets dropped here, together with any other values
            // created in the critical section.
        };
        // The guard created here is a temporary dropped at the end of the statement, i.e.
        // the lock would not remain being held even if the thread did some additional work.
        *res_mutex_clone.lock() += result;
    }));
});

let mut data = data_mutex.lock();
// This is the result of some important and long-ish work.
let result = data.iter().fold(0, |acc, x| acc + x * 2);
data.push(result);
// We drop the `data` explicitly because it's not necessary anymore and the
// thread still has work to do. This allows other threads to start working on
// the data immediately, without waiting for the rest of the unrelated work
// to be done here.
//
// It's even more important here than in the threads because we `.join` the
// threads after that. If we had not dropped the mutex guard, a thread could
// be waiting forever for it, causing a deadlock.
// As in the threads, a block could have been used instead of calling the
// `drop` function.
drop(data);
// Here the mutex guard is not assigned to a variable and so, even if the
// scope does not end after this line, the mutex is still released: there is
// no deadlock.
*res_mutex.lock() += result;

threads
    .into_iter()
    .for_each(|thread| thread.join().expect("The thread creating or execution failed !"));

assert_eq!(*res_mutex.lock(), 800);

Implementations

impl<T> Mutex<T>

pub const fn new(t: T) -> Mutex<T>

Creates a new mutex in an unlocked state ready for use.

Examples

use pspsdk::sync::poison::Mutex;

let mutex = Mutex::new(0);

impl<T, M: RawMutex> Mutex<T, M>

pub const fn new_with(value: T, raw_mutex: M) -> Self

Creates a new mutex in an unlocked state ready for use using the specified mutex.

pub fn get_cloned(&self) -> Result<T, PoisonError<()>>
where T: Clone,

Returns the contained value by cloning it.

Errors

If another user of this mutex panicked while holding the mutex, then this call will return an error instead.

Examples

use pspsdk::sync::poison::Mutex;

let mut mutex = Mutex::new(7);

assert_eq!(mutex.get_cloned().unwrap(), 7);

pub fn set(&self, value: T) -> Result<(), PoisonError<T>>

Sets the contained value.

Errors

If another user of this mutex panicked while holding the mutex, then this call will return an error containing the provided value instead.

Examples

use pspsdk::sync::poison::Mutex;

let mut mutex = Mutex::new(7);

assert_eq!(mutex.get_cloned().unwrap(), 7);
mutex.set(11).unwrap();
assert_eq!(mutex.get_cloned().unwrap(), 11);

pub fn replace(&self, value: T) -> LockResult<T>

Replaces the contained value with value, and returns the old contained value.

Errors

If another user of this mutex panicked while holding the mutex, then this call will return an error containing the provided value instead.

Examples

use pspsdk::sync::poison::Mutex;

let mut mutex = Mutex::new(7);

assert_eq!(mutex.replace(11).unwrap(), 7);
assert_eq!(mutex.get_cloned().unwrap(), 11);

impl<T: ?Sized, M: RawMutex> Mutex<T, M>

pub fn lock(&self) -> MutexGuard<'_, T, M>

Acquires a mutex, blocking the current thread until it is able to do so.

This function will block the local thread until it is available to acquire the mutex. Upon returning, the thread is the only thread with the lock held. An RAII guard is returned to allow scoped unlock of the lock. When the guard goes out of scope, the mutex will be unlocked.

The exact behavior on locking a mutex in the thread which already holds the lock is left unspecified. However, this function will not return on the second call (it might panic or deadlock, for example).

Panics

This function might panic when called if the lock is already held by the current thread or the lock is in a poisoned state.

Examples

use std::{sync::Arc, thread};

use pspsdk::sync::poison::Mutex;

let mutex = Arc::new(Mutex::new(0));
let c_mutex = Arc::clone(&mutex);

thread::spawn(move || {
    *c_mutex.lock() = 10;
})
.join()
.expect("thread::spawn failed");
assert_eq!(*mutex.lock(), 10);

pub fn try_lock(&self) -> TryLockResult<MutexGuard<'_, T, M>>

Attempts to acquire this lock.

If the lock could not be acquired at this time, then [Err] is returned. Otherwise, an RAII guard is returned. The lock will be unlocked when the guard is dropped.

This function does not block.

Errors

If another user of this mutex panicked while holding the mutex, then this call will return the Poisoned error if the mutex would otherwise be acquired. An acquired lock guard will be contained in the returned error.

If the mutex could not be acquired because it is already locked, then this call will return the WouldBlock error.

Examples

use std::{sync::Arc, thread};

use pspsdk::sync::poison::Mutex;

let mutex = Arc::new(Mutex::new(0));
let c_mutex = Arc::clone(&mutex);

thread::spawn(move || {
    let mut lock = c_mutex.try_lock();
    if let Ok(ref mut mutex) = lock {
        **mutex = 10;
    } else {
        println!("try_lock failed");
    }
})
.join()
.expect("thread::spawn failed");
assert_eq!(*mutex.lock(), 10);

pub fn lock_checked(&self) -> LockResult<MutexGuard<'_, T, M>>

Acquires a mutex, blocking the current thread until it is able to do so.

This function will block the local thread until it is available to acquire the mutex. Upon returning, the thread is the only thread with the lock held. An RAII guard is returned to allow scoped unlock of the lock. When the guard goes out of scope, the mutex will be unlocked.

The exact behavior on locking a mutex in the thread which already holds the lock is left unspecified. However, this function will not return on the second call (it might panic or deadlock, for example).

Errors

If another user of this mutex panicked while holding the mutex, then this call will return an error once the mutex is acquired. The acquired mutex guard will be contained in the returned error.

Panics

This function might panic when called if the lock is already held by the current thread.

Examples

use std::{sync::Arc, thread};

use pspsdk::sync::poison::Mutex;

let mutex = Arc::new(Mutex::new(0));
let c_mutex = Arc::clone(&mutex);

thread::spawn(move || {
    *c_mutex.lock_checked().unwrap() = 10;
})
.join()
.expect("thread::spawn failed");
assert_eq!(*mutex.lock(), 10);

pub fn lock_unchecked(&self) -> MutexGuard<'_, T, M>

Same as lock, but ignoring poison state instead of panicking.

Panics

This function might panic when called if the lock is already held by the current thread.

pub fn try_unchecked(&self) -> TryLockResult<MutexGuard<'_, T, M>>

Same as try_lock, but ignoring poison state.

Errors

If the mutex could not be acquired because it is already locked, then this call will return the WouldBlock error.

pub fn is_poisoned(&self) -> bool

Determines whether the mutex is poisoned.

If another thread is active, the mutex can still become poisoned at any time. You should not trust a false value for program correctness without additional synchronization.

Examples

use std::{sync::Arc, thread};

use pspsdk::sync::poison::Mutex;

let mutex = Arc::new(Mutex::new(0));
let c_mutex = Arc::clone(&mutex);

let _ = thread::spawn(move || {
    let _lock = c_mutex.lock();
    panic!(); // the mutex gets poisoned
})
.join();
assert_eq!(mutex.is_poisoned(), true);

pub fn clear_poison(&self)

Clear the poisoned state from a mutex.

If the mutex is poisoned, it will remain poisoned until this function is called. This allows recovering from a poisoned state and marking that it has recovered. For example, if the value is overwritten by a known-good value, then the mutex can be marked as un-poisoned. Or possibly, the value could be inspected to determine if it is in a consistent state, and if so the poison is removed.

Examples

use std::{sync::Arc, thread};

use pspsdk::sync::poison::Mutex;

let mutex = Arc::new(Mutex::new(0));
let c_mutex = Arc::clone(&mutex);

let _ = thread::spawn(move || {
    let _lock = c_mutex.lock();
    panic!(); // the mutex gets poisoned
})
.join();

assert_eq!(mutex.is_poisoned(), true);
let x = mutex.lock_checked().unwrap_or_else(|mut e| {
    **e.get_mut() = 1;
    mutex.clear_poison();
    e.into_inner()
});
assert_eq!(mutex.is_poisoned(), false);
assert_eq!(*x, 1);

pub fn into_inner(self) -> T
where T: Sized,

Consumes this mutex, returning the underlying data.

Panics

If another user of this mutex panicked while holding the mutex, then this call will panic instead.

Examples

use pspsdk::sync::poison::Mutex;

let mutex = Mutex::new(0);
assert_eq!(mutex.into_inner(), 0);

pub fn into_inner_checked(self) -> LockResult<T>
where T: Sized,

Consumes this mutex, returning the underlying data, explicitly cheking poison state.

Errors

If another user of this mutex panicked while holding the mutex, then this call will return an error containing the the underlying data instead.

Examples

use pspsdk::sync::poison::Mutex;

let mutex = Mutex::new(0);
assert_eq!(mutex.into_inner_checked().unwrap(), 0);

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data.

Since this call borrows the Mutex mutably, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist.

Panics

If another user of this mutex panicked while holding the mutex, then this call will panic instead.

Examples

use pspsdk::sync::poison::Mutex;

let mut mutex = Mutex::new(0);
*mutex.get_mut() = 10;
assert_eq!(*mutex.lock(), 10);

pub fn get_mut_checked(&mut self) -> LockResult<&mut T>

Returns a mutable reference to the underlying data, explicitly cheking poison state.

Since this call borrows the Mutex mutably, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist.

Errors

If another user of this mutex panicked while holding the mutex, then this call will return an error containing a mutable reference to the underlying data instead.

Examples

use pspsdk::sync::poison::Mutex;

let mut mutex = Mutex::new(0);
*mutex.get_mut_checked().unwrap() = 10;
assert_eq!(*mutex.lock(), 10);

pub fn into_inner_unchecked(self) -> T
where T: Sized,

Consumes this mutex ignoring the poison state, returning the underlying data.

Examples

use pspsdk::sync::poison::Mutex;

let mutex = Mutex::new(0);
assert_eq!(mutex.into_inner_unchecked(), 0);

pub fn get_mut_unchecked(&mut self) -> &mut T

Returns a mutable reference to the underlying data, ignoring poison state.

Since this call borrows the Mutex mutably, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist.

Examples

use pspsdk::sync::poison::Mutex;

let mut mutex = Mutex::new(0);
*mutex.get_mut_unchecked() = 10;
assert_eq!(*mutex.lock(), 10);

pub const fn data_ptr(&self) -> *mut T

Returns a raw pointer to the underlying data.

The returned pointer is always non-null and properly aligned, but it is the user’s responsibility to ensure that any reads and writes through it are properly synchronized to avoid data races, and that it is not read or written through after the mutex is dropped.

impl<T: ?Sized, M: RawMutexTimed> Mutex<T, M>

pub fn try_lock_for(&self, timeout: Duration) -> Option<MutexGuard<'_, T, M>>

Attempts to acquire this lock until a timeout is reached.

If the lock could not be acquired before the timeout expired, then None is returned. Otherwise, an RAII guard is returned. The lock will be unlocked when the guard is dropped.

pub fn try_lock_until(&self, timeout: Instant) -> Option<MutexGuard<'_, T, M>>

Attempts to acquire this lock until a timeout is reached.

If the lock could not be acquired before the timeout expired, then None is returned. Otherwise, an RAII guard is returned. The lock will be unlocked when the guard is dropped.