pspsdk::sync::poison

Struct PoisonError

pub struct PoisonError<T> { /* private fields */ }

Available on crate feature non-stub-code only.

A type of error which can be returned whenever a lock is acquired.

Both Mutexes and RwLocks are poisoned whenever a thread fails while the lock is held. The precise semantics for when a lock is poisoned is documented on each lock. For a lock in the poisoned state, unless the state is cleared manually, all future acquisitions will return this error.

Examples

use std::{
    sync::{Arc, Mutex},
    thread,
};

let mutex = Arc::new(Mutex::new(1));

// poison the mutex
let c_mutex = Arc::clone(&mutex);
let _ = thread::spawn(move || {
    let mut data = c_mutex.lock().unwrap();
    *data = 2;
    panic!();
})
.join();

match mutex.lock() {
    Ok(_) => unreachable!(),
    Err(p_err) => {
        let data = p_err.get_ref();
        println!("recovered: {data}");
    },
};

Implementations

impl<T> PoisonError<T>

pub fn new(data: T) -> PoisonError<T>

Available on panic=unwind only.

Creates a PoisonError.

This is generally created by methods like Mutex::lock_checked or RwLock::read_checked.

This method may panic if std was built with panic="abort".

pub fn into_inner(self) -> T

Consumes this error indicating that a lock is poisoned, returning the underlying guard to allow access regardless.

Examples

use std::{
    collections::HashSet,
    sync::{Arc, Mutex},
    thread,
};

let mutex = Arc::new(Mutex::new(HashSet::new()));

// poison the mutex
let c_mutex = Arc::clone(&mutex);
let _ = thread::spawn(move || {
    let mut data = c_mutex.lock().unwrap();
    data.insert(10);
    panic!();
})
.join();

let p_err = mutex.lock().unwrap_err();
let data = p_err.into_inner();
println!("recovered {} items", data.len());

pub fn get_ref(&self) -> &T

Reaches into this error indicating that a lock is poisoned, returning a reference to the underlying guard to allow access regardless.

pub fn get_mut(&mut self) -> &mut T

Reaches into this error indicating that a lock is poisoned, returning a mutable reference to the underlying guard to allow access regardless.