pspsdk::sync::poison

Struct RwLock

pub struct RwLock<T: ?Sized, Raw = SpinRwLock> { /* private fields */ }

Available on crate feature non-stub-code only.

A reader-writer lock

This type of lock allows a number of readers or at most one writer at any point in time. The write portion of this lock typically allows modification of the underlying data (exclusive access) and the read portion of this lock typically allows for read-only access (shared access).

In comparison, a Mutex does not distinguish between readers or writers that acquire the lock, therefore blocking any threads waiting for the lock to become available. An RwLock will allow any number of readers to acquire the lock as long as a writer is not holding the lock.

The priority policy of the lock is dependent on the underlying operating system’s implementation, and this type does not guarantee that any particular policy will be used. In particular, a writer which is waiting to acquire the lock in write might or might not block concurrent calls to read, e.g.:

Potential deadlock example

// Thread 1              |  // Thread 2
let _rg1 = lock.read();  |
                         |  // will block
                         |  let _wg = lock.write();
// may deadlock          |
let _rg2 = lock.read();  |

The type parameter T represents the data that this lock protects. It is required that T satisfies [Send] to be shared across threads and [Sync] to allow concurrent access through readers. The RAII guards returned from the locking methods implement [Deref] (and [DerefMut] for the write methods) to allow access to the content of the lock.

Poisoning

An RwLock, like Mutex, will become poisoned on a panic. Note, however, that an RwLock may only be poisoned if a panic occurs while it is locked exclusively (write mode). If a panic occurs in any reader, then the lock will not be poisoned.

Examples

use pspsdk::sync::poison::RwLock;

let lock = RwLock::new(5);

// many reader locks can be held at once
{
    let r1 = lock.read();
    let r2 = lock.read();
    assert_eq!(*r1, 5);
    assert_eq!(*r2, 5);
} // read locks are dropped at this point

// only one write lock may be held, however
{
    let mut w = lock.write();
    *w += 1;
    assert_eq!(*w, 6);
} // write lock is dropped here

Implementations

impl<T> RwLock<T>

pub const fn new(t: T) -> RwLock<T>

Creates a new instance of an RwLock<T> which is unlocked.

Examples

use pspsdk::sync::poison::RwLock;

let lock = RwLock::new(5);

impl<T, L: RawRwLock> RwLock<T, L>

pub const fn new_with(value: T, raw_lock: L) -> Self

Creates a new instance of an RwLock<T> which is unlocked using the specified lock.

pub fn get_cloned(&self) -> Result<T, PoisonError<()>>
where T: Clone,

Returns the contained value by cloning it.

Errors

This function will return an error if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock.

Examples

use pspsdk::sync::poison::RwLock;

let mut lock = RwLock::new(7);

assert_eq!(lock.get_cloned().unwrap(), 7);

pub fn set(&self, value: T) -> Result<(), PoisonError<T>>

Sets the contained value.

Errors

This function will return an error containing the provided value if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock.

Examples

use pspsdk::sync::poison::RwLock;

let mut lock = RwLock::new(7);

assert_eq!(lock.get_cloned().unwrap(), 7);
lock.set(11).unwrap();
assert_eq!(lock.get_cloned().unwrap(), 11);

pub fn replace(&self, value: T) -> LockResult<T>

Replaces the contained value with value, and returns the old contained value.

Errors

This function will return an error containing the provided value if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock.

Examples

use pspsdk::sync::poison::RwLock;

let mut lock = RwLock::new(7);

assert_eq!(lock.replace(11).unwrap(), 7);
assert_eq!(lock.get_cloned().unwrap(), 11);

impl<T: ?Sized, L: RawRwLock> RwLock<T, L>

pub fn read(&self) -> RwLockReadGuard<'_, T, L>

Locks this RwLock with shared read access, blocking the current thread until it can be acquired.

The calling thread will be blocked until there are no more writers which hold the lock. There may be other readers currently inside the lock when this method returns. This method does not provide any guarantees with respect to the ordering of whether contentious readers or writers will acquire the lock first.

Returns an RAII guard which will release this thread’s shared access once it is dropped.

Panics

This function might panic when called if the lock is already held by the current thread.

This function will panic if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock. The failure will occur immediately after the lock has been acquired.

Examples

use std::{sync::Arc, thread};

use pspsdk::sync::poison::RwLock;

let lock = Arc::new(RwLock::new(1));
let c_lock = Arc::clone(&lock);

let n = lock.read();
assert_eq!(*n, 1);

thread::spawn(move || {
    let r = c_lock.read();
    assert_eq!(*r, 1);
})
.join()
.unwrap();

pub fn try_read(&self) -> TryLockResult<RwLockReadGuard<'_, T, L>>

Attempts to acquire this RwLock with shared read access.

If the access could not be granted at this time, then Err is returned. Otherwise, an RAII guard is returned which will release the shared access when it is dropped.

This function does not block.

This function does not provide any guarantees with respect to the ordering of whether contentious readers or writers will acquire the lock first.

Errors

This function will return the Poisoned error if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock. Poisoned will only be returned if the lock would have otherwise been acquired. An acquired lock guard will be contained in the returned error.

This function will return the WouldBlock error if the RwLock could not be acquired because it was already locked exclusively.

Examples

use pspsdk::sync::poison::RwLock;

let lock = RwLock::new(1);

match lock.try_read() {
    Ok(n) => assert_eq!(*n, 1),
    Err(_) => unreachable!(),
};

pub fn write(&self) -> RwLockWriteGuard<'_, T, L>

Locks this RwLock with exclusive write access, blocking the current thread until it can be acquired.

This function will not return while other writers or other readers currently have access to the lock.

Returns an RAII guard which will drop the write access of this RwLock when dropped.

Panics

This function might panic when called if the lock is already held by the current thread.

This function will panic if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock. An error will be returned when the lock is acquired.

Examples

use pspsdk::sync::poison::RwLock;

let lock = RwLock::new(1);

let mut n = lock.write();
*n = 2;

assert!(lock.try_read().is_err());

pub fn try_write(&self) -> TryLockResult<RwLockWriteGuard<'_, T, L>>

Attempts to lock this RwLock with exclusive write access.

If the lock could not be acquired at this time, then Err is returned. Otherwise, an RAII guard is returned which will release the lock when it is dropped.

This function does not block.

This function does not provide any guarantees with respect to the ordering of whether contentious readers or writers will acquire the lock first.

Errors

This function will return the Poisoned error if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock. Poisoned will only be returned if the lock would have otherwise been acquired. An acquired lock guard will be contained in the returned error.

This function will return the WouldBlock error if the RwLock could not be acquired because it was already locked exclusively.

Examples

use pspsdk::sync::poison::RwLock;

let lock = RwLock::new(1);

let n = lock.read();
assert_eq!(*n, 1);

assert!(lock.try_write().is_err());

pub fn read_unchecked(&self) -> RwLockReadGuard<'_, T, L>

Locks this RwLock with shared read access ignoring poison state, blocking the current thread until it can be acquired.

The calling thread will be blocked until there are no more writers which hold the lock. There may be other readers currently inside the lock when this method returns. This method does not provide any guarantees with respect to the ordering of whether contentious readers or writers will acquire the lock first.

Returns an RAII guard which will release this thread’s shared access once it is dropped.

Panics

This function might panic when called if the lock is already held by the current thread.

Examples

use std::{sync::Arc, thread};

use pspsdk::sync::poison::RwLock;

let lock = Arc::new(RwLock::new(1));
let c_lock = Arc::clone(&lock);

let n = lock.read();
assert_eq!(*n, 1);

thread::spawn(move || {
    let r = c_lock.read_unchecked();
    assert_eq!(*r, 1);
})
.join()
.unwrap();

pub fn write_unchecked(&self) -> RwLockWriteGuard<'_, T, L>

Locks this RwLock with exclusive write access ignoring poison state, blocking the current thread until it can be acquired.

This function will not return while other writers or other readers currently have access to the lock.

Returns an RAII guard which will drop the write access of this RwLock when dropped.

Panics

This function might panic when called if the lock is already held by the current thread.

This function will panic if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock. An error will be returned when the lock is acquired.

Examples

use pspsdk::sync::poison::RwLock;

let lock = RwLock::new(1);

let mut n = lock.write();
*n = 2;

assert!(lock.try_read().is_err());

pub fn into_inner(self) -> T
where T: Sized,

Consumes this RwLock, returning the underlying data.

Panics

This function will panic if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock.

Examples

use pspsdk::sync::poison::RwLock;

let lock = RwLock::new(String::new());
{
    let mut s = lock.write();
    *s = "modified".to_owned();
}
assert_eq!(lock.into_inner(), "modified");

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data.

Since this call borrows the RwLock mutably, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist.

Panics

This function will panic if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock.

Examples

use pspsdk::sync::poison::RwLock;

let mut lock = RwLock::new(0);
*lock.get_mut() = 10;
assert_eq!(*lock.read(), 10);

pub fn into_inner_unchecked(self) -> T
where T: Sized,

Consumes this RwLock, returning the underlying data without checks if the lock is poisoned.

pub fn get_mut_unchecked(&mut self) -> &mut T

Returns a mutable reference to the underlying data without checks if the lock is poisoned.

Since this call borrows the RwLock mutably, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist.

impl<T: ?Sized, L: RawRwLock> RwLock<T, L>

pub fn read_checked(&self) -> LockResult<RwLockReadGuard<'_, T, L>>

Locks this RwLock with shared read access, blocking the current thread until it can be acquired.

The calling thread will be blocked until there are no more writers which hold the lock. There may be other readers currently inside the lock when this method returns. This method does not provide any guarantees with respect to the ordering of whether contentious readers or writers will acquire the lock first.

Returns an RAII guard which will release this thread’s shared access once it is dropped.

Errors

This function will return an error if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock. The failure will occur immediately after the lock has been acquired. The acquired lock guard will be contained in the returned error.

Panics

This function might panic when called if the lock is already held by the current thread.

Examples

use std::{sync::Arc, thread};

use pspsdk::sync::poison::RwLock;

let lock = Arc::new(RwLock::new(1));
let c_lock = Arc::clone(&lock);

let n = lock.read();
assert_eq!(*n, 1);

thread::spawn(move || {
    let r = c_lock.read_checked();
    assert!(r.is_ok());
})
.join()
.unwrap();

pub fn write_checked(&self) -> LockResult<RwLockWriteGuard<'_, T, L>>

Locks this RwLock with exclusive write access, blocking the current thread until it can be acquired.

This function will not return while other writers or other readers currently have access to the lock.

Returns an RAII guard which will drop the write access of this RwLock when dropped.

Errors

This function will return an error if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock. An error will be returned when the lock is acquired. The acquired lock guard will be contained in the returned error.

Panics

This function might panic when called if the lock is already held by the current thread.

Examples

use pspsdk::sync::poison::RwLock;

let lock = RwLock::new(1);

let mut n = lock.write();
*n = 2;

assert!(lock.try_read().is_err());

pub fn is_poisoned(&self) -> bool

Determines whether the lock is poisoned.

If another thread is active, the lock can still become poisoned at any time. You should not trust a false value for program correctness without additional synchronization.

Examples

use std::{
    sync::{Arc, RwLock},
    thread,
};

let lock = Arc::new(RwLock::new(0));
let c_lock = Arc::clone(&lock);

let _ = thread::spawn(move || {
    let _lock = c_lock.write().unwrap();
    panic!(); // the lock gets poisoned
})
.join();
assert_eq!(lock.is_poisoned(), true);

pub fn clear_poison(&self)

Clear the poisoned state from a lock.

If the lock is poisoned, it will remain poisoned until this function is called. This allows recovering from a poisoned state and marking that it has recovered. For example, if the value is overwritten by a known-good value, then the lock can be marked as un-poisoned. Or possibly, the value could be inspected to determine if it is in a consistent state, and if so the poison is removed.

Examples

use std::{
    sync::{Arc, RwLock},
    thread,
};

let lock = Arc::new(RwLock::new(0));
let c_lock = Arc::clone(&lock);

let _ = thread::spawn(move || {
    let _lock = c_lock.write().unwrap();
    panic!(); // the lock gets poisoned
})
.join();

assert_eq!(lock.is_poisoned(), true);
let guard = lock.write().unwrap_or_else(|mut e| {
    **e.get_mut() = 1;
    lock.clear_poison();
    e.into_inner()
});
assert_eq!(lock.is_poisoned(), false);
assert_eq!(*guard, 1);

pub fn into_inner_checked(self) -> LockResult<T>
where T: Sized,

Consumes this RwLock, returning the underlying data.

Errors

This function will return an error containing the underlying data if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock. An error will only be returned if the lock would have otherwise been acquired.

Examples

use pspsdk::sync::poison::RwLock;

let lock = RwLock::new(String::new());
{
    let mut s = lock.write();
    *s = "modified".to_owned();
}
assert_eq!(lock.into_inner_checked().unwrap(), "modified");

pub fn get_mut_checked(&mut self) -> LockResult<&mut T>

Returns a mutable reference to the underlying data.

Since this call borrows the RwLock mutably, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist.

Errors

This function will return an error containing a mutable reference to the underlying data if the RwLock is poisoned. An RwLock is poisoned whenever a writer panics while holding an exclusive lock. An error will only be returned if the lock would have otherwise been acquired.

Examples

use pspsdk::sync::poison::RwLock;

let mut lock = RwLock::new(0);
*lock.get_mut_checked().unwrap() = 10;
assert_eq!(*lock.read(), 10);

pub const fn data_ptr(&self) -> *mut T

Returns a raw pointer to the underlying data.

The returned pointer is always non-null and properly aligned, but it is the user’s responsibility to ensure that any reads and writes through it are properly synchronized to avoid data races, and that it is not read or written through after the mutex is dropped.

impl<T: ?Sized, L: RawRwLockTimed> RwLock<T, L>

pub fn try_read_for( &self, timeout: Duration, ) -> Option<RwLockReadGuard<'_, T, L>>

Attempts to acquire this RwLock with shared read access until a timeout is reached.

If the access could not be granted before the timeout expires, then None is returned. Otherwise, an RAII guard is returned which will release the shared access when it is dropped.

pub fn try_read_until( &self, timeout: Instant, ) -> Option<RwLockReadGuard<'_, T, L>>

Attempts to acquire this RwLock with shared read access until a timeout is reached.

If the access could not be granted before the timeout expires, then None is returned. Otherwise, an RAII guard is returned which will release the shared access when it is dropped.

pub fn try_write_for( &self, timeout: Duration, ) -> Option<RwLockWriteGuard<'_, T, L>>

Attempts to acquire this RwLock with exclusive write access until a timeout is reached.

If the access could not be granted before the timeout expires, then None is returned. Otherwise, an RAII guard is returned which will release the exclusive access when it is dropped.

pub fn try_write_until( &self, timeout: Instant, ) -> Option<RwLockWriteGuard<'_, T, L>>

Attempts to acquire this RwLock with exclusive write access until a timeout is reached.

If the access could not be granted before the timeout expires, then None is returned. Otherwise, an RAII guard is returned which will release the exclusive access when it is dropped.