pspsdk::sync

Trait RawMutex

pub trait RawMutex: Sealed {
    const NEW: Self;

    // Required methods
    fn lock(&self);
    fn try_lock(&self) -> bool;
    unsafe fn unlock(&self);

    // Provided method
    fn is_locked(&self) -> bool { ... }
}

Available on crate feature non-stub-code only.

Define basic operations for a mutex.

A type that implements this can be used with the Mutex type to have a mutex that functions similarly as the Rust standard library mutex.

Required Associated Constants

const NEW: Self

A constant new to overcome the limitations of traits not having const fn yet on trits

Required Methods

fn lock(&self)

Locks the mutex or panics if the lock is already held by the current thread.

fn try_lock(&self) -> bool

Tries to acquire this mutex without blocking, returning true if the lock was successfull and false otherwise.

unsafe fn unlock(&self)

Unlocks this mutex.

Safety

This method may only be called if the mutex is held in the current context, i.e. it must be paired with a successful call to lock, try_lock, try_lock_for or try_lock_until.

Provided Methods

fn is_locked(&self) -> bool

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".