pspsdk::sync

Trait RawMutexTimed

pub trait RawMutexTimed: RawMutex {
    // Required method
    fn try_lock_for(&self, timeout: Duration) -> bool;

    // Provided method
    fn try_lock_until(&self, timeout: Instant) -> bool { ... }
}

Available on crate feature non-stub-code only.

Define timed operation on a mutex.

Required Methods

fn try_lock_for(&self, timeout: Duration) -> bool

Attempts to acquire this lock for a timeout duration.

Provided Methods

fn try_lock_until(&self, timeout: Instant) -> bool

Attempts to acquire this lock until a timeout is reached.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".