pspsdk::sync

Trait RawRwLock

pub trait RawRwLock: Sealed {
    const NEW: Self;

    // Required methods
    fn read(&self);
    fn try_read(&self) -> bool;
    fn write(&self);
    fn try_write(&self) -> bool;
    unsafe fn read_unlock(&self);
    unsafe fn write_unlock(&self);
    unsafe fn downgrade(&self);

    // Provided methods
    fn is_locked(&self) -> bool { ... }
    fn is_locked_exclusive(&self) -> bool { ... }
}

Available on crate feature non-stub-code only.

Define basic operations for a ReadWrite Lock.

A type that implements this can be used with the RwLock type to have a lock that functions similarly as the Rust standard library rwlock.

Required Associated Constants

const NEW: Self

Required Methods

fn read(&self)

Acquires a read lock, blocking the current thread until it is able to do so.

fn try_read(&self) -> bool

Attempts to acquire a read lock without blocking.

fn write(&self)

Acquires an write lock, blocking the current thread until it is able to do so.

fn try_write(&self) -> bool

Attempts to acquire an write lock without blocking.

unsafe fn read_unlock(&self)

Releases a read lock.

Safety

This method may only be called if a read lock is held in the current context.

unsafe fn write_unlock(&self)

Releases an write lock.

Safety

This method may only be called if an write lock is held in the current context.

unsafe fn downgrade(&self)

Atomically downgrades an write lock into a read lock without allowing any thread to take an write lock in the meantime.

Safety

This method may only be called if an write lock is held in the current context.

Provided Methods

fn is_locked(&self) -> bool

Checks if this RwLock is currently locked in any way.

fn is_locked_exclusive(&self) -> bool

Check if this RwLock is currently exclusively locked (write locked).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".