pspsdk::sync

Trait RawRwLockTimed

pub trait RawRwLockTimed: RawRwLock {
    // Required methods
    fn try_read_for(&self, timeout: Duration) -> bool;
    fn try_write_for(&self, timeout: Duration) -> bool;

    // Provided methods
    fn try_read_until(&self, timeout: Instant) -> bool { ... }
    fn try_write_until(&self, timeout: Instant) -> bool { ... }
}

Available on crate feature non-stub-code only.

Required Methods

fn try_read_for(&self, timeout: Duration) -> bool

Attempts to acquire a read lock for a timeout duration.

fn try_write_for(&self, timeout: Duration) -> bool

Attempts to acquire a write lock for a timeout duration.

Provided Methods

fn try_read_until(&self, timeout: Instant) -> bool

Attempts to acquire a read lock until a timeout is reached.

fn try_write_until(&self, timeout: Instant) -> bool

Attempts to acquire a write lock until a timeout is reached.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".