pspsdk

Module sys

The low-level implementations, type definition, and system stubs for the PSP system.

Modules

audio

Audio operation and management.

ctrl

Input control operations.

display

Display operations.

dma

Direct Memory Access (DMA) operations.

ge

Graphic Engine operation and management.

hprm

Headphone Remote control.

io

Input/Output operations.

libc

Libc functions.

library

Libraries information and types.

loadexec

Load, execution and exit management.

mem

Memory management.

module

Module management.

openpsid

PS ID operations.

power

Power management.

suspend

Suspend operations.

syncnon-stub-code

Sync primitives implementation.

thread

Thread operations and management.

time

Time operation and management.

usersystemlib

The usersustemlib.prx exported functions.

Structs

LibFlags

Resident/Stub library attributes.

SceError

Type representing return errors from the PSP APIs, official or otherwise.

SceResult

A type that represents the return value of many PSP OS APIs.

SceResult64

A type that represents the return value of some PSP OS APIs.

SceUid

Identification number for several kernel objects.

Enums

ErrorFacility

The source facility of an error value.

Traits

SceInto64OkValue

Trait that specifies that it can safely be turned to a SceResult64 ok-valid value.

SceIntoOkValue

Trait that specifies that it can safely be turned to a SceResult ok-valid value.

SceResultOk

Trait of types that can be the ok result of SceResult or SceResult64.

Functions

disable_fpu_exceptionsPlay Station Portable and non-stub-code

Disables the CPU FPU exceptions.

get_current_interrupt_statusPlay Station Portable and non-stub-code

Disables interrupts, returning the previous state of the interrupt enable bit.

get_k1Play Station Portable and non-stub-code

Gets the current value of the processor K1 register.

is_interrupt_enabledPlay Station Portable and non-stub-code

Returns true if the the interrupt is enabled, false otherwise.

resume_interrupts⚠Play Station Portable and non-stub-code

Enables interrupts to the state value.

set_k1⚠Play Station Portable and non-stub-code

Sets the processor K1 register to a given value.

spin_loopPlay Station Portable and non-stub-code

Intrinsic to hint a spin-loop to the mips processor

suspend_interrupts⚠Play Station Portable and non-stub-code

Disables interrupts, returning the previous state of the interrupt enable bit.

volatile_read⚠non-stub-code

Performs a volatile read of the value from addr without moving it. This leaves the memory in addr unchanged.

volatile_write⚠non-stub-code

Performs a volatile write of a memory location addr with the given value without reading or dropping the old value.

Type Aliases

SceIsize

A isize-like with the guarantee to be the correct size on PSP.

SceRawUid

Identification number for several kernel objects.

SceSize

A usize-like with the guarantee to be the correct size on PSP.