pspsdk::sys

Module audio

Audio operation and management.

Modules

atrac

Atrac format operation and processing.

codec

Codec operations.

routing

Audio routing control.

Structs

AudioChannelId

Representation of the PSP channel number.

AudioInputParams

Enums

AudioChannels

Possible values for number of audio channels.

AudioFormats

Possible audio formats for PSP.

AudioInputCompletionStatus

The audio input completion status returned by sceAudioPollInputEnd.

AudioInputFrequency

Possible values for Audio input frequency.

AudioOutputFrequency

Possible values for Audio output frequency.

Constants

AUDIO_SAMPLE_MAX

Maximum value for audio sample value.

AUDIO_SAMPLE_MIN

Minimum value for audio sample value.

Functions

audio_sample_align

Make the given sample count a multiple of 64.

sceAudioChRelease⚠Play Station Portable

Releases an initialized audio channel ID.

sceAudioChReservePlay Station Portable

Allocate and initialize a hardware output channel.

sceAudioChangeChannelConfig⚠Play Station Portable

Changes the audio format of a channel.

sceAudioChangeChannelVolume⚠Play Station Portable

Changes the volume of a channel.

sceAudioGetChannelRestLen⚠Play Station Portable

Gets the count of unplayed samples remaining.

sceAudioGetChannelRestLength⚠Play Station Portable

Gets the count of unplayed samples remaining.

sceAudioGetInputLengthPlay Station Portable

Gets the number of samples that were acquired.

sceAudioInput⚠Play Station Portable

Performs audio input asynchronously.

sceAudioInputBlocking⚠Play Station Portable

Performs audio input (blocking).

sceAudioInputInitDeprecatedPlay Station Portable

Initializes the audio input.

sceAudioInputInitExPlay Station Portable

Initializes the audio input.

sceAudioOutput⚠Play Station Portable

Gets the output audio of a channel asynchronously.

sceAudioOutput2ChangeLengthPlay Station Portable

Change the output sample count after it’s already been reserved.

sceAudioOutput2GetRestSamplePlay Station Portable

Gets the count of unplayed samples remaining.

sceAudioOutput2OutputBlockingPlay Station Portable

Gets the output audio of a channel (blocking).

sceAudioOutput2ReleasePlay Station Portable

Releases the audio output and process termination.

sceAudioOutput2ReservePlay Station Portable

Reserves the audio output and set the output sample count.

sceAudioOutputBlocking⚠Play Station Portable

Gets the output audio of a channel (blocking).

sceAudioOutputPanned⚠Play Station Portable

Gets the panned output audio of a channel asynchronously.

sceAudioOutputPannedBlocking⚠Play Station Portable

Gets the panned output audio of a channel (blocking).

sceAudioPollInputEndPlay Station Portable

Polls the completion status for non-blocking audio input.

sceAudioSRCChReleasePlay Station Portable

Releases the audio output.

sceAudioSRCChReservePlay Station Portable

Reserves the audio output.

sceAudioSetChannelDataLen⚠Play Station Portable

Sets the output sample count, after it’s already been reserved.

sceAudioWaitInputEndPlay Station Portable

Waits for a non-blocking audio input to complete.