pspsdk::sys::audio

Module atrac

Atrac format operation and processing.

Structs

AtracBufferInfo

AtracId

Identification of Atrac3 objects.

Enums

AtracCodecKind

The possible kinds of ATRAC codec.

AtracLoopStatus

The looping status of the Atrac object.

Functions

sceAtracAddStreamDataPlay Station Portable

Adds a specified number of bytes to the atrac stream data.

sceAtracDecodeData⚠Play Station Portable

Decode a frame of data.

sceAtracGetAtracIDPlay Station Portable

Get the Atrac ID for an available/released Atrac object with the specified codec_kind.

sceAtracGetBitratePlay Station Portable

Gets the Atrac object bitrate.

sceAtracGetBufferInfoForResetingPlay Station Portable

Gets the buffer information to reset the Atrac buffer.

sceAtracGetChannelPlay Station Portable

Get the channel of a Atrac object.

sceAtracGetInternalErrorInfoPlay Station Portable

Get the internal Codec error from the Atrac object.

sceAtracGetLoopStatusPlay Station Portable

Get the loop status and loop iteration from the Atrac object.

sceAtracGetMaxSamplePlay Station Portable

Gets the maximum number of samples of the Atrac3 stream.

sceAtracGetNextDecodePositionPlay Station Portable

Gets the next decode sample position.

sceAtracGetNextSamplePlay Station Portable

Gets the number of samples of the next frame to be decoded.

sceAtracGetRemainFramePlay Station Portable

Gets the remaining (not decoded) number of frames

sceAtracGetSecondBufferInfoPlay Station Portable

Get the second buffer information of the Atrac object.

sceAtracGetSoundSamplePlay Station Portable

Get the sound sample information of the Atrac object.

sceAtracGetStreamDataInfoPlay Station Portable

Get information of the Atrac stream data of the given Atrac ID.

sceAtracReleaseAtracIDPlay Station Portable

Releases an Atrac ID

sceAtracResetPlayPositionPlay Station Portable

Resets the play position of the sample of a Atrac object.

sceAtracSetData⚠Play Station Portable

sceAtracSetDataAndGetID⚠Play Station Portable

Creates a new Atrac ID from the specified data.

sceAtracSetHalfwayBuffer⚠Play Station Portable

sceAtracSetHalfwayBufferAndGetID⚠Play Station Portable

sceAtracSetLoopNumPlay Station Portable

Sets the number of loops for the Atrac object associated to the given Atrac ID.

sceAtracSetSecondBuffer⚠Play Station Portable