pspsdk::sys::audio::routing

Enum AudioRoutingMode

#[repr(u32)]
pub enum AudioRoutingMode {
    Auto = 0,
    PreferSpeakers = 1,
    OnlySpeakers = 2,
    OnlyHeadphone = 3,
}

The routing mode behavior.

Variants

Auto = 0

Audio output on speaker, automatically change to headphone when it is plugged.

PreferSpeakers = 1

Audio output on speaker, do not automatically change to headphone when it is plugged.

OnlySpeakers = 2

Audio output only on speakers, always.

OnlyHeadphone = 3

Audio output only on headphone, always.