pspsdk::sys

Module ctrl

Input control operations.

Structs

ButtonEmulationSlot

The slot for a button emulation setup.

CtrlData

Structure used to obtain input data from the controller.

CtrlDataExt

Structure used to obtain input extended data from the controller.

IdleCancelThreshold

A valid idle cancel threshold value.

InputDataTransferHandle

Type used to copy external input data into PSP internal controller buffers.

LatchData

Represents a controller latch data.

PadButtons

Flags representing digital controller button.

RapidFireSlot

The slot for a rapid-fire setup.

SamplingCycle

A valid sampling cycle value in microseconds.

SpecialButtonCallbackSlot

The slot for a special button callback setup.

Enums

ButtonMaskMode

Button mask setting that specifies the type of input data to be obtained.

ControllerPort

Specifies the kind of input data to be obtained.

InputMode

Controller input sampling mode.

Functions

sceCtrlClearRapidFirePlay Station Portable

Disables a rapid-fire button event.

sceCtrlGetButtonInterceptPlay Station Portable

Gets the button mask settings applied to PSP buttons.

sceCtrlGetIdleCancelKeyPlay Station Portable

Gets the different cancel-idle-timer buttons.

sceCtrlGetIdleCancelThresholdPlay Station Portable

Gets the currently set idle timer cancel movement threshold values for the analog stick.

sceCtrlGetSamplingCyclePlay Station Portable

Gets the currently set interval of the internal controller data buffers.

sceCtrlGetSamplingModePlay Station Portable

Gets the currently set input sampling mode.

sceCtrlGetSuspendingExtraSamplesPlay Station Portable

Gets the number of VBlanks which will be waited for when the PSP device is being suspended.

sceCtrlPeekBufferNegative⚠Play Station Portable

Retrieves controller state data by polling (negative logic, i.e. state data will be 0 when the button is pressed).

sceCtrlPeekBufferNegative2⚠Play Station Portable

Retrieves controller state data on a specified controller port by polling (negative logic, i.e. state data will be 0 when the button is pressed).

sceCtrlPeekBufferPositive⚠Play Station Portable

Retrieves controller state data by polling (positive logic, i.e. state data will be 1 when the button is pressed).

sceCtrlPeekBufferPositive2⚠Play Station Portable

Retrieves controller state data on a specified controller port by polling (positive logic, i.e. state data will be 1 when the button is pressed).

sceCtrlPeekLatch⚠Play Station Portable

Gets the latest latch data from the controller service.

sceCtrlReadBufferNegative⚠Play Station Portable

Retrieves controller state data by blocking (negative logic, i.e. state data will be 0 when the button is pressed).

sceCtrlReadBufferNegative2⚠Play Station Portable

Retrieves controller state data on a specified controller port by blocking (negative logic, i.e. state data will be 0 when the button is pressed).

sceCtrlReadBufferPositive⚠Play Station Portable

Retrieves controller state data by blocking (positive logic, i.e. state data will be 1 when the button is pressed).

sceCtrlReadBufferPositive2⚠Play Station Portable

Retrieves controller state data on a specified controller port by blocking (positive logic, i.e. state data will be 1 when the button is pressed).

sceCtrlReadLatch⚠Play Station Portable

Reads new latch data from the controller service.

sceCtrlSetAnalogEmulationPlay Station Portable

Emulates values for the analog pad’s X- and Y-axis.

sceCtrlSetButtonEmulationPlay Station Portable

Emulates buttons for the digital pad.

sceCtrlSetButtonInterceptPlay Station Portable

Sets a button mask mode for one or more buttons.

sceCtrlSetIdleCancelKeyPlay Station Portable

Sets the buttons which, when being pressed, reset the idle timer.

sceCtrlSetIdleCancelThresholdPlay Station Portable

Sets analog stick movement threshold values for cancelling the idle timer.

sceCtrlSetInternalBufferHandler⚠Play Station Portable

Sets up internal controller buffers to receive external input data.

sceCtrlSetRapidFirePlay Station Portable

Sets a rapid-fire event for one or more buttons.

sceCtrlSetSamplingCyclePlay Station Portable

Sets the update frequency of the internal controller buffer.

sceCtrlSetSamplingModePlay Station Portable

Sets the controller input sampling mode.

sceCtrlSetSpecialButtonCallbackPlay Station Portable

Registers a button callback.

sceCtrlSetSuspendingExtraSamplesPlay Station Portable

Sets a number of VBlanks which will be waited for when the PSP device is being suspended.

Type Aliases

InputDataTransferFn

Represents the function pointer of InputDataTransferHandle.

SpecialButtonCallback

The callback function used by sceCtrlSetSpecialButtonCallback.