pspsdk::sys::ctrl

Function sceCtrlGetIdleCancelThreshold

pub extern "C" fn sceCtrlGetIdleCancelThreshold(
    unhold_threshold: &mut IdleCancelThreshold,
    hold_threshold: &mut IdleCancelThreshold,
) -> SceResult<()>

Available on Play Station Portable only.

Gets the currently set idle timer cancel movement threshold values for the analog stick.

Parameters

unhold_threshold: A reference to receive the movement needed by the analog stick to reset the idle timer when HOLD mode is inactive.

hold_threshold: A reference to receive the movement needed by the analog stick to reset the idle timer when HOLD mode is active.

Return Value

Ok value on success, error value otherwise.

Firmware Version

This API was introduced on PSP firmware version 1.50.