pspsdk::sys::ctrl

Function sceCtrlSetRapidFire

pub extern "C" fn sceCtrlSetRapidFire(
    slot: RapidFireSlot,
    mask: PadButtons,
    trigger: PadButtons,
    target: PadButtons,
    delay: u32,
    make: u32,
    breakk: u32,
) -> SceResult<()>

Available on Play Station Portable only.

Sets a rapid-fire event for one or more buttons.

Parameters

slot: The slot of the event to set.

mask: The comparison mask of the button operation for rapid-fire trigger. In order for the trigger buttons to trigger the event, they need to be included in these buttons.

trigger: The buttons which will start the rapid fire event for the specified target buttons when being pressed.

target: The buttons for which the rapid-fire event will be applied to. User mode buttons only. make and breakk define the rapid-fire cycle.

delay: The dead time of rapid-fire trigger (sampling count). Specifies the rapid-fire start timing. It will only be applied for the first ON period of a (not cancelled) rapid-fire event. Valid value in the range of [0,63].

make: The press time for the target buttons. This “ON-time” is set after delay was applied and the trigger buttons were turned OFF. It will be applied for as long as the same rapid fire event is called without a break (i.e. pressing of a different PSP button). Valid value in the range of [0,63]. If set to 0, the target button(s) will be turned ON for one sampling count.

breakk: The release time for target buttons. This “OFF-time” is set after delay was applied. It will be applied as long as the same rapid fire event is called without a break (i.e. the pressing of a different PSP button). Valid value in the range of [0,63]. If set to 0, the target button will be turned OFF for 64 consecutive sampling counts.

Return Value

Ok value on success, error value otherwise.