pspsdk::sys::ctrl

Struct IdleCancelThreshold

pub struct IdleCancelThreshold(/* private fields */);

A valid idle cancel threshold value.

Implementations

impl IdleCancelThreshold

pub const ALWAYS_CANCEL: Self

When idle cancel threshold is set to this value, the idle timer is canceled even if the analog stick is not moved (i.e., the idle timer itself stops running).

pub const NO_CANCEL: Self

When idle cancel threshold is set to this value, no cancellation the idle timer happens no matter the analog stick movement.

pub const fn new(raw: i32) -> Result<Self, SceError>

Creates a new idle cancel threshold if it is one of the valid values.

pub const unsafe fn new_unchecked(raw: i32) -> Self

Creates a new idle cancel threshold without checking for it’s validity.

Safety

The value must be 0 or in the range of [5555,20000] (or, alternatively [180Hz, 50Hz]).

pub const fn to_inner(self) -> i32

Gets the inner value of the idle cancel threshold value.