pspsdk::sys::ctrl

Struct PadButtons

pub struct PadButtons(/* private fields */)
where
    u32: BitsPrimitive;

Flags representing digital controller button.

Each flag corresponds to a different button and can be used to extract button states from CtrlData and LatchData structures. Flags can be combined using bitwise OR operation to check for multiple key states at once.

Note

The following buttons can only be read at a kernel-level: PadButtons::Home, PadButtons::WlanUp, PadButtons::Remote, PadButtons::VolumeUp, PadButtons::VolumeDown, PadButtons::Screen, PadButtons::Note, PadButtons::Disc, and PadButtons::MemoryStick

Implementations

impl PadButtons

pub const Select: Self

Select button.

pub const L3: Self

L3 button.

pub const R3: Self

R3 button.

pub const Start: Self

Start button.

pub const Up: Self

Up D-Pad button.

pub const Right: Self

Right D-Pad button.

pub const Down: Self

Down D-Pad button.

pub const Left: Self

Left D-Pad button.

pub const LTrigger: Self

Left trigger.

This accounts for the L2 trigger as well.

pub const RTrigger: Self

Right trigger.

This accounts for the R2 trigger as well.

pub const L1Trigger: Self

L1 trigger.

pub const R1Trigger: Self

R1 trigger.

pub const L2Trigger: Self

L2 trigger.

pub const R2Trigger: Self

R2 trigger.

pub const Triangle: Self

Triangle button.

pub const Circle: Self

Circle button.

pub const Cross: Self

Cross button.

pub const Square: Self

Square button.

pub const Intercepted: Self

If this bit is set, then controller input is being intercepted by the system software or another application. For example, this is the case when the PSP’s HOME menu or the exit dialog is being shown.

pub const Home: Self

Home button.

pub const Hold: Self

Hold button.

pub const WlanUp: Self

W-LAN switch up.

pub const Remote: Self

Remote hold position.

pub const VolumeUp: Self

Volume up button.

pub const VolumeDown: Self

Volume down button.

pub const Screen: Self

Screen button.

pub const Note: Self

Music Note button.

pub const Disc: Self

Disc is present.

pub const MemoryStick: Self

Memory Stick is present.

pub const Unk10000000: Self

pub const Unk20000000: Self

impl PadButtons

pub const fn bits(&self) -> u32

Return the underlying bits value.

pub const fn from_bits(bits: u32) -> Option<Self>

Converts from a bits value. Returning [None] is any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> Self

Convert from bits value exactly.

pub fn from_flag_name(name: &str) -> Option<Self>

Convert from a flag name.

pub const fn empty() -> Self

Construct a flags value with all bits unset.

pub const fn is_empty(&self) -> bool

Returns true if the flags value has all bits unset.

pub const fn all_bits() -> Self

Returns a flags value that contains all value.

This will include bits that do not have any flags/meaning. Use all if you want only the specified flags set.

pub const fn is_all_bits(&self) -> bool

Returns true if the flags value contains all value bits set.

This will check for all bits. Use is_all if you want to check for all specified flags.

pub const fn all() -> Self

Construct a flags value with all known flags set.

This will only set the flags specified as associated constant and the defined extra valid bits.

pub const fn is_all(&self) -> bool

Returns true if the flags value contais all known flags.

pub const fn all_named() -> Self

Construct a flags value with all known named flags set.

This will only set the flags specified as associated constant without the defined extra valid bits.

pub const fn is_all_named(&self) -> bool

Returns true if the flags value contais all known named flags.

pub const fn contains_unknown_bits(&self) -> bool

Returns true if there are any unknown bits set in the flags value.

pub const fn contains_unnamed_bits(&self) -> bool

Returns true if there are any unnamed known bits set in the flags value.

pub const fn truncated(&self) -> Self

Returns a flags value with unknown bits removed from the original flags value.

pub const fn truncate(&mut self)

Removes unknown bits from the flags value.

pub const fn intersects(&self, other: Self) -> bool

Returns true if this flags value intersects with any value in other.

This is equivalent to (self & other) != Self::empty()

pub const fn contains(&self, other: Self) -> bool

Returns true if this flags value contains all values of other.

This is equivalent to (self & other) == other

pub const fn not(self) -> Self

Returns the bitwise NOT of the flags value.

This function does not truncate unused bits (bits that do not have any flags/meaning). Use complement if you want that the result to be truncated in one call.

pub const fn and(self, other: Self) -> Self

Returns the bitwise AND of the flags value with other.

pub const fn or(self, other: Self) -> Self

Returns the bitwise OR of the flags value with other.

pub const fn xor(self, other: Self) -> Self

Returns the bitwise XOR of the flags value with other.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection from this flags value with other.

pub const fn union(self, other: Self) -> Self

Returns the union from this flags value with other.

pub const fn difference(self, other: Self) -> Self

Returns the difference from this flags value with other.

In other words, returns the intersection of this flags value with the negation of other.

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference from this flags value with other.

pub const fn complement(self) -> Self

Returns the complement of the flags value.

This is very similar to the not, but truncates non used bits.

pub const fn set(&mut self, other: Self)

Set the flags in other in the flags value.

pub const fn unset(&mut self, other: Self)

Unset the flags bits in other in the flags value.

pub const fn toggle(&mut self, other: Self)

Toggle the flags in other in the flags value.

pub const fn clear(&mut self)

Resets the flags value to a empty state.

impl PadButtons

pub const fn iter(&self) -> Iter<Self>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<Self>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.