pspsdk::sys::ctrl

Struct RapidFireSlot

pub struct RapidFireSlot(/* private fields */);

The slot for a rapid-fire setup.

Implementations

impl RapidFireSlot

pub const SLOT_00: Self

Rapid fire slot #00.

pub const SLOT_01: Self

Rapid fire slot #01.

pub const SLOT_02: Self

Rapid fire slot #02.

pub const SLOT_03: Self

Rapid fire slot #03.

pub const SLOT_04: Self

Rapid fire slot #04.

pub const SLOT_05: Self

Rapid fire slot #05.

pub const SLOT_06: Self

Rapid fire slot #06.

pub const SLOT_07: Self

Rapid fire slot #07.

pub const SLOT_08: Self

Rapid fire slot #08.

pub const SLOT_09: Self

Rapid fire slot #09.

pub const SLOT_10: Self

Rapid fire slot #10.

pub const SLOT_11: Self

Rapid fire slot #11.

pub const SLOT_12: Self

Rapid fire slot #12.

pub const SLOT_13: Self

Rapid fire slot #13.

pub const SLOT_14: Self

Rapid fire slot #14.

pub const SLOT_15: Self

Rapid fire slot #15.

pub const fn new(raw: u32) -> Option<Self>

Create a new RapidFireSlot structure from a raw value.

This functions checks for the value of raw to be a in the range of possible SceUid values used by the PSP OS, returning an [None] otherwise.

pub const unsafe fn new_unchecked(raw: u32) -> Self

Create a new RapidFireSlot structure from a raw value without checking value range.

Safety

Immediate language UB if val is not within the valid range for this type, as it violates the validity invariant.

pub const fn to_inner(self) -> u32