pspsdk::sys

Module display

Display operations.

Enums

DisplayHoldMode

The display mode during a soft reboot (*LoadExec* family of functions).

DisplayMode

Display output modes.

DisplayState

The display state.

DisplayUpdateSync

Display change sync strategies.

PixelFormat

Functions

sceDisplayDisablePlay Station Portable

Disables the display.

sceDisplayEnablePlay Station Portable

Enables the display.

sceDisplayEnd⚠Play Station Portable

Deinitilizes the sceDisplay* library.

sceDisplayGetAccumulatedHcountPlay Station Portable

Gets the accumulated HSYNC count.

sceDisplayGetBacklightSelPlay Station Portable

Gets the display backlight brightness.

sceDisplayGetBrightnessPlay Station Portable

Gets the current brightness level set.

sceDisplayGetCurrentHcountPlay Station Portable

Gets the current HSYNC count.

sceDisplayGetFrameBuf⚠Play Station Portable

Gets the display framebuffer.

sceDisplayGetFramePerSecPlay Station Portable

Gets the number of frames per second in the current display mode.

sceDisplayGetModePlay Station Portable

Gets the current display mode and display proportions.

sceDisplayGetVcountPlay Station Portable

Gets the number of vertical blank pulses up to now

sceDisplayInit⚠Play Station Portable

Initializes the sceDisplay* library.

sceDisplayIsForegroundPlay Station Portable

Gets if the framebuffer is current being displayed.

sceDisplayIsVblankPlay Station Portable

Checks whether VBLANK is active in process.

sceDisplaySetBacklightSelPlay Station Portable

Sets the display backlight brightness to a particular level.

sceDisplaySetBrightnessPlay Station Portable

Sets the display brightness to a particular level.

sceDisplaySetFrameBuf⚠Play Station Portable

Sets the display framebuffer.

sceDisplaySetHoldModePlay Station Portable

Sets the hold mode.

sceDisplaySetModePlay Station Portable

Sets the display mode.

sceDisplayWaitVblankPlay Station Portable

Makes the calling thread to enter in a Wait state until VBLANK is in process.

sceDisplayWaitVblankCBPlay Station Portable

Makes the calling thread to enter in a Wait state until VBLANK is in process, but service any callbacks as necessary.

sceDisplayWaitVblankStartPlay Station Portable

Makes the calling thread to enter in a Wait state until the the start of the next VBLANK.

sceDisplayWaitVblankStartCBPlay Station Portable

Makes the calling thread to enter in a Wait state until the the start of the next VBLANK, but service any callbacks as necessary.

sceDisplayWaitVblankStartMultiPlay Station Portable

Makes the calling thread to enter in a Wait state until the the start of the next VBLANK after a number of VSYNC cycles.

sceDisplayWaitVblankStartMultiCBPlay Station Portable

Makes the calling thread to enter in a Wait state until the the start of the next VBLANK after a number of VSYNC cycles, but service any callbacks as necessary.