pspsdk::sys

Function volatile_read

pub unsafe fn volatile_read<T>(addr: usize) -> T
where
    T: VolatileOpAllowed,

Available on crate feature non-stub-code only.

Performs a volatile read of the value from addr without moving it. This leaves the memory in addr unchanged.

Volatile operations are intended to act on I/O memory, and are guaranteed to not be elided or reordered by the compiler across other volatile operations.

Safety

Behavior is undefined if any of the following conditions are violated:

addr must be either valid for reads, or addr must point to memory outside of all Rust allocations and reading from that memory must:

not trap, and

not cause any memory inside a Rust allocation to be modified.

addr must be properly aligned.