pspsdk::sys

Module ge

Graphic Engine operation and management.

Structs

BreakParam

Break unused parameter.

DisplayListId

A Valid Display list ID.

GeBreakpoint

A Graphics Engine breakpoint.

GeCallbackId

A Graphic Engine callback ID.

GeCallbackOptions

The Graphics Engine callback options.

GeContext

Storage for a Graphic Engine context.

GeListArgs

The list of arguments when enqueueing a list.

GeStack

Storage for a stack (for CALL/RET).

Enums

BreakMode

The break mode to interrupt display list operation.

DisplayListState

The state of a display list.

DisplayListSyncMode

Graphics Engine display list synchronization mode.

EdramSizeMode

Size mode for the EDRAM.

GeCmd

The available GE commands.

MatrixRegister

The registers of the Graphic Engine matrices.

TranslationMemWidth

The possible width values of the EDRAM address translation.

Functions

sceGeBreak⚠Play Station Portable

Interrupts the Graphic Engine drawing.

sceGeContinuePlay Station Portable

Continues the queue execution after a sceGeBreak.

sceGeDrawSyncPlay Station Portable

Synchronizes a display list drawing.

sceGeEdramGetAddrPlay Station Portable

Gets the physical eDRAM address.

sceGeEdramGetHwSizePlay Station Portable

Gets the EDRAM physical size.

sceGeEdramGetSizePlay Station Portable

Gets the size of eDRAM.

sceGeEdramSetAddrTranslationPlay Station Portable

Sets the EDRAM address translation.

sceGeEdramSetSizePlay Station Portable

Sets the EDRAM size to be enabled.

sceGeGetBreakpoint⚠Play Station Portable

Gets set breakpoints.

sceGeGetCmdPlay Station Portable

Get the value of a command.

sceGeGetListIdList⚠Play Station Portable

Gets a list of the IDs of the display lists currently being in the queue.

sceGeGetMtx⚠Play Station Portable

Get a matrix from a matrix register.

sceGeListDeQueuePlay Station Portable

Dequeues a display list.

sceGeListEnQueuePlay Station Portable

Enqueues a display list at the end of the queue.

sceGeListEnQueueHeadPlay Station Portable

Enqueues a display list at the start of the queue.

sceGeListSyncPlay Station Portable

Synchronizes a display list.

sceGeListUpdateStallAddr⚠Play Station Portable

Updates the stall address of a display list.

sceGePutBreakpoint⚠Play Station Portable

Put breakpoints in the display list execution.

sceGeRestoreContextPlay Station Portable

Restores the Graphics Engine internal state context.

sceGeSaveContextPlay Station Portable

Saves the Graphics Engine internal state context of the current display list.

sceGeSetCallbackPlay Station Portable

Sets Graphics Engine finish/signal callbacks.

sceGeSetMtx⚠Play Station Portable

Sets a matrix of a matrix register.

sceGeUnsetCallbackPlay Station Portable

Unset a Graphics Engine finish/signal callbacks.

Type Aliases

GeCallbackFn

Represents the function pointer of GeCallbackOptions.