pspsdk::sys

Module io

Input/Output operations.

Structs

AssignFlag

Mount flags used on sceIoAssign.

ChangeStatFlag

What field of stat to change on sceIoChstat.

Cwd

Current working directory.

DeviceAttribute

Device attributes.

DeviceEntry

Structure passed to the drive function DriveFunctionTable::init and DriveFunctionTable::exit.

Dirent

A single directory entry.

DirentFatPrivate

Private information of a directory (specific to FAT filesystem, which is the only one the PSP has support anyway).

DriveDevice

Drive device information to add with sceIoAddDrv.

DriveFunctionTable

Type to maintain the device driver pointers.

FileFlags

Flags that control how a file is opened and its subsequent behavior.

FileId

File descriptor UID.

IoBlockInfo

Device information.

IoHook

Mode

Typed representation of file modes, combining file type and permissions.

Stat

The status information of a file.

StatAttribute

Attributes of the file Stat.

Enums

DeviceKind

Possible device kinds.

Whence

Possible methods to seek within an file object.

Functions

sceIoAddDrv⚠Play Station Portable

Adds a new I/O driver to the system.

sceIoAssign⚠Play Station Portable

Assigns one I/O device to another.

sceIoCancelPlay Station Portable

Cancels a file asynchronous task.

sceIoChangeAsyncPriorityPlay Station Portable

Changes the priority level of a file asynchronous task.

sceIoChangeThreadCwd⚠Play Station Portable

Sets the current working directory for a thread.

sceIoChdir⚠Play Station Portable

Changes the current directory.

sceIoChstat⚠Play Station Portable

Changes the status of a file.

sceIoClose⚠Play Station Portable

Closes a open file and deletes the associated file descriptor ID.

sceIoCloseAsync⚠Play Station Portable

Closes a open file and deletes the associated file descriptor ID asynchronously.

sceIoDclose⚠Play Station Portable

Closes an directory opened with sceIoDopen.

sceIoDelDrv⚠Play Station Portable

Removes a I/O driver from the system.

sceIoDevctl⚠Play Station Portable

Performs an Device Control command on a device.

sceIoDopen⚠Play Station Portable

Opens a directory.

sceIoDreadPlay Station Portable

Reads an entry from a opened file descriptor UID associated to a file opened with sceIoDopen.

sceIoGetAsyncStatPlay Station Portable

Gets the completions status of a file asynchronous task.

sceIoGetDevTypePlay Station Portable

Gets the device attribute of the opened file descriptor.

sceIoGetFdList⚠Play Station Portable

Gets the list of opened file descriptor UIDs.

sceIoGetIobUserLevelPlay Station Portable

Gets the user level out of I/O block.

sceIoGetThreadCwd⚠Play Station Portable

Gets the current working directory for a thread.

sceIoGetUIDPlay Station Portable

Gets the real file descriptor UID from global list of UID.

sceIoGetstat⚠Play Station Portable

Gets the status of a file.

sceIoIoctl⚠Play Station Portable

Performs an I/O Control command on a file.

sceIoIoctlAsync⚠Play Station Portable

Performs an I/O Control command on a file asynchronously.

sceIoLseekPlay Station Portable

Repositions the file offset of the file descriptor.

sceIoLseek32Play Station Portable

Repositions the file offset of the file descriptor. (32-bit mode).

sceIoLseek32AsyncPlay Station Portable

Repositions the file offset of the file descriptor asynchronously. (32-bit mode).

sceIoLseekAsyncPlay Station Portable

Repositions the file offset of the file descriptor asynchronously.

sceIoMkdir⚠Play Station Portable

Creates a directory file.

sceIoOpen⚠Play Station Portable

Opens a file with the corresponding file access attributes.

sceIoOpenAsync⚠Play Station Portable

Opens a file with the corresponding file access attributes asynchronously.

sceIoPollAsyncPlay Station Portable

Polls a file asynchronous task completion.

sceIoRead⚠Play Station Portable

Reads data from the file associated to the file descriptor UID to a buffer.

sceIoReadAsync⚠Play Station Portable

Reads data from the file associated to the file descriptor UID to a buffer asynchronously.

sceIoRemove⚠Play Station Portable

Removes a file associated to a given path.

sceIoRename⚠Play Station Portable

Changes the file name to a new name.

sceIoReopen⚠Play Station Portable

Reopens an existing file descriptor UID.

sceIoRmdir⚠Play Station Portable

Removes a directory file associated to a given path.

sceIoSetAsyncCallback⚠Play Station Portable

Sets a callback for the file asynchronous task.

sceIoSync⚠Play Station Portable

Synchronize the file data on the device.

sceIoTerminateFd⚠Play Station Portable

Forcedly removes a I/O driver from the system.

sceIoUnassign⚠Play Station Portable

Unassign an I/O device.

sceIoValidateFdPlay Station Portable

Checks for the validity of a file descriptor.

sceIoWaitAsyncPlay Station Portable

Waits for a file asynchronous action completion.

sceIoWaitAsyncCBPlay Station Portable

Waits for a file asynchronous action completion with callbacks.

sceIoWrite⚠Play Station Portable

Writes data to the file associated to the file descriptor UID.

sceIoWriteAsync⚠Play Station Portable

Writes data to the file associated to the file descriptor UID asynchronously.

sceKernelRegisterStderrPipePlay Station Portable

Register a file descriptor as PIPE to stderr.

sceKernelRegisterStdoutPipePlay Station Portable

Register a file descriptor as PIPE to stdout.

sceKernelStderrPlay Station Portable

Function to get the current standard error file ID

sceKernelStderrReopen⚠Play Station Portable

Reopens the standard Error.

sceKernelStderrResetPlay Station Portable

Resets the standard Error.

sceKernelStdinPlay Station Portable

Function to get the current standard in file ID

sceKernelStdoutPlay Station Portable

Function to get the current standard out file ID

sceKernelStdoutReopen⚠Play Station Portable

Reopens the standard Out.

sceKernelStdoutResetPlay Station Portable

Resets the standard Out.