pspsdk::sys::io

Function sceIoOpenAsync

pub unsafe extern "C" fn sceIoOpenAsync(
    path: *const u8,
    flags: FileFlags,
    mode: Mode,
) -> SceResult<FileId>

Available on Play Station Portable only.

Opens a file with the corresponding file access attributes asynchronously.

The function only supports full path.

If the flag FileFlags::CreateFile is set in flags, the third parameter is used if the file needs to be created, otherwise it is ignored.

Parameters

path [[In parameter]]: The path to the file to open.

flags: The flags with file access attributes.

mode: The permission mode to use. If FileFlags::CreateFile is set in flags, it is ignored otherwise.

Result Value

Returns the open file descriptor ID on success, error value otherwise.