pspsdk::sys::io

Struct DeviceEntry

#[repr(C)]
pub struct DeviceEntry {
    pub drv: *mut DriveDevice,
    pub private_data: *mut c_void,
    pub user_fd_count: SceSize,
}

Structure passed to the drive function DriveFunctionTable::init and DriveFunctionTable::exit.

Fields
drv: *mut DriveDevice

A pointer to the original driver which was added.

private_data: *mut c_void

A pointer to a setter defined argument.

If written by the driver will preserve across calls.

The actual data types used is device dependent.

user_fd_count: SceSize

The number of file descriptor UID related to the device.