pspsdk::sys::io

Struct IoBlockInfo

#[repr(C)]
pub struct IoBlockInfo {

    pub flags: FileFlags,
    pub fs_unit: u32,
    pub dev_entry: *mut DeviceEntry,
    pub dev_attr: DeviceAttribute,
    pub private_data: *mut c_void,
    pub cwd: *mut Cwd,
    pub file_pos: i64,
    pub thread_id: ThreadId,
    pub unk: u32,
    pub uid: FileId,
    pub open_thread_id: ThreadId,
    pub is_user_mode: bool,
    pub is_power_locked: bool,
    pub unk2: u8,
    pub async_priority: u8,
    pub async_thread_id: ThreadId,
    pub async_sema: SemaId,
    pub async_event_flag: EventFlagId,
    pub async_callback: CallbackId,
    pub async_callback_argp: *mut c_void,
    pub unused: u32,
    pub k1: u32,
    pub async_return: i64,
    pub async_args: [i32; 6],
    pub async_cmd: u32,
    pub user_level: u32,
    pub hook: IoHook,
    pub unk3: u32,
    pub new_path: *mut u8,
    pub return_addr: SceSize,

}

Device information.

Mainly used for the parameter for the DriveFunctionTable function pointers.

Fields
flags: FileFlags

The file flags

fs_unit: u32

The file system number.

E.g.: If a file is opened as host5:/myfile.txt this field will be 5.

dev_entry: *mut DeviceEntry

A pointer to the device entry data.

This pointer should never be null in the context of DriveFunctionTable functions being called, but it can be before that.

dev_attr: DeviceAttribute

The device attribute.

It contains the device kind.

private_data: *mut c_void

A pointer to a setter defined argument.

If written by the driver will preserve across calls.

The actual data types used is device dependent.

cwd: *mut Cwdfile_pos: i64

The current file position offset.

thread_id: ThreadId

The thread UID of (?).

unk: u32

Currently unknown

uid: FileId

Raw UID of the opened file

open_thread_id: ThreadId

The thread UID of the thread that opened the file.

is_user_mode: bool

If the file is opened in user mode.

is_power_locked: bool

If the file was opened with FileFlags::PowerLock.

unk2: u8

Unknown. It is a boolean of some sort, async related.

async_priority: u8

The async task priority

async_thread_id: ThreadId

The thread UID of the thread for the file async operations.

async_sema: SemaId

The semaphore UID for the file async operations.

async_event_flag: EventFlagId

The event flag UID for the file async operations.

async_callback: CallbackId

The callback UID for the file async operations.

async_callback_argp: *mut c_void

A pointer to the parameter to pass to async_callback.

unused: u32

Unused.

k1: u32

The current k1 register value for async operations

async_return: i64

The async return value of operations.

async_args: [i32; 6]

Arguments for the command to use.

async_cmd: u32

The async command to execute.

user_level: u32

The current user level of

hook: IoHookunk3: u32new_path: *mut u8return_addr: SceSize

The return address.