pspsdk::sys

Module libc

Libc functions.

PSP has many sources for those functions:

SysclibForKernel - kernel only

scePaf - VSH/XMB only

SysclibForUser - CFW lib

PSPSDK - static lib, here they are substituted to a Rust implementation

Whenever possible, using one of the first three is ideal to reuse the implementation and reduce the software size and memory usage.

NOTE: SysclibForUser export those functions as syscalls, that means they are slightly less performant than SysclibForKernel and scePaf in the user-level context.

Functions

__udivdi3⚠Play Station Portable

__udivmoddi4⚠Play Station Portable

__umoddi3⚠Play Station Portable

bcmp⚠Play Station Portable

bcopy⚠Play Station Portable

bzero⚠Play Station Portable

index⚠Play Station Portable

memchr⚠Play Station Portable

memcmp⚠Play Station Portable

memcpy⚠Play Station Portable

memmove⚠Play Station Portable

memset⚠Play Station Portable

rindex⚠Play Station Portable

snprintf⚠Play Station Portable

sprintf⚠Play Station Portable

strcat⚠Play Station Portable

strchr⚠Play Station Portable

strcmp⚠Play Station Portable

strcpy⚠Play Station Portable

strlen⚠Play Station Portable

strncmp⚠Play Station Portable

strncpy⚠Play Station Portable

strnlen⚠Play Station Portable

strpbrk⚠Play Station Portable

strrchr⚠Play Station Portable

strstr⚠Play Station Portable

strtol⚠Play Station Portable

strtoul⚠Play Station Portable

tolower⚠Play Station Portable

toupper⚠Play Station Portable

wmemset⚠Play Station Portable