pspsdk::sys

Module library

Libraries information and types.

Re-exports

pub use crate::sys::module::ModuleAttributes;

Structs

FunctionDirectCall

This type represents a function stub belonging to the same privilege-level linked libraries, i.e. a kernel resident library linked with a kernel stub library.

FunctionSyscall

This type represents a function stub belonging to different privilege-level linked libraries, i.e. a kernel resident library linked with a user stub library.

ModuleInfo

The information of a module.

ResidentLibraryEntry

Represents a single resident export library of a module.

StubLibraryEntry

Represents the imports, provided by a resident library, that a given module is using.

VariableStub

This type represents an imported variable stub.

Constants

RESI_LIBRARY_ENTRY_TABLE_NEW_LEN

RESI_LIBRARY_ENTRY_TABLE_OLD_LEN

STUB_LIBRARY_ENTRY_TABLE_NEW_LEN

STUB_LIBRARY_ENTRY_TABLE_OLD_LEN

Unions

FunctionStub

ResidentLibraryEntryItem