pspsdk::sys::library

Struct ModuleInfo

#[repr(C)]
pub struct ModuleInfo {
    pub attributes: ModuleAttributes,
    pub version: (u8, u8),
    pub name: [u8; 27],
    pub terminal_char: u8,
    pub gp: *const u8,
    pub entry_top: *const u8,
    pub entry_end: *const u8,
    pub stub_top: *const u8,
    pub stub_end: *const u8,
}

The information of a module.

Fields
attributes: ModuleAttributes

The attributes of a module.

version: (u8, u8)

The version of the module.

It consists of a ‘major’ and ‘minor’ field.

name: [u8; 27]

The name of the module.

terminal_char: u8

The string terminator.

It is always b'\0'.

gp: *const u8

The global pointer of the module.

entry_top: *const u8

A pointer to the first resident library entry table of the module.

This section is known as “.lib.ent”.

entry_end: *const u8

A pointer to the last line of the “.lib.ent” section.

This line is always 0 and it is known as “.lib.ent.btm”.

stub_top: *const u8

A pointer to the first stub library entry table of the module.

This section is known as “lib.stub”.

stub_end: *const u8

A pointer to the last line of the “lib.stub” section.

This line is always `0`` and is known as “.lib.stub.btm”.