pspsdk::sys::library

Struct StubLibraryEntry

#[repr(C)]
pub struct StubLibraryEntry {
    pub name: *const u8,
    pub version: (u8, u8),
    pub flags: LibFlags,
    pub len: u8,
    pub var_stub_count: u8,
    pub func_stub_count: u16,
    pub nid_table: *const u32,
    pub func_stub_table: *const FunctionStub,
    pub var_stub_table: *const VariableStub,
    pub unk: u32,
}

Represents the imports, provided by a resident library, that a given module is using.

A module can have multiple stub libraries.

Fields
name: *const u8

The name of the library.

version: (u8, u8)

The version of the library.

It consists of a ‘major’ and ‘minor’ field.

The version of a stub library shouldn’t be higher than the version(s) of the corresponding resident library/libraries. Linking won’t be performed in such a case.

flags: LibFlags

The library’s flags.

It can be set to either LibFlags::NoSpecialFlags or LibFlags::WeakImport.

len: u8

The length of the entry table in 32-Bit words.

Set this to either STUB_LIBRARY_ENTRY_TABLE_OLD_LEN or STUB_LIBRARY_ENTRY_TABLE_NEW_LEN.

Use this member when you want to iterate through a list of entry tables (size = len * 4).

var_stub_count: u8

The number of imported variables by the stub library.

func_stub_count: u16

The number of imported functions by the stub library.

nid_table: *const u32

A pointer to an array of NIDs containing the NIDs of the imported functions and variables.

func_stub_table: *const FunctionStub

A pointer to an array of imported function stubs.

var_stub_table: *const VariableStub

A pointer to an array of imported variable stubs.

unk: u32