pspsdk::sys

Module loadexec

Load, execution and exit management.

Structs

LoadExecOptions

Options for sceKernelLoadExec.

LoadExecVshOptions

Options to configure the sceKernelVSH* functions behavior.

Functions

sceKernelCheckExitCallbackPlay Station Portable

Gets the current exit callback.

sceKernelExitGamePlay Station Portable

Exits the game and reboot back to the XMB/VSH.

sceKernelExitGameWithStatusPlay Station Portable

Exits the game and reboot back to the XMB/VSH with a status value.

sceKernelExitVSHKernelPlay Station Portable

Restart the VSH (to be used by a kernel module).

sceKernelExitVSHVSHPlay Station Portable

Restarts the VSH.

sceKernelInvokeExitCallbackPlay Station Portable

Invokes the exit callback.

sceKernelLoadExec⚠Play Station Portable

Load and execute a new executable.

sceKernelLoadExecBufferPlain⚠Play Station Portable

Executes a new executable from a buffer.

sceKernelLoadExecVSHDisc⚠Play Station Portable

Executes a executable from a disc.

sceKernelLoadExecVSHDiscDebug⚠Play Station Portable

Executes a debug executable from a disc.

sceKernelLoadExecVSHDiscUpdater⚠Play Station Portable

Executes a updater executable from a disc.

sceKernelLoadExecVSHMs1⚠Play Station Portable

Executes a updater executable from a memory stick.

sceKernelLoadExecVSHMs2⚠Play Station Portable

Executes a executable from a memory stick.

sceKernelLoadExecVSHMs3⚠Play Station Portable

Executes a executable from a memory stick.

sceKernelLoadExecVSHMs4⚠Play Station Portable

Executes a App executable from a memory stick.

sceKernelLoadExecVSHMs5⚠Play Station Portable

Executes a PS1 executable from a memory stick.

sceKernelRegisterExitCallbackPlay Station Portable

Registers a callback that is executed when “Exit Game” is confirmed in the home button menu.

sceKernelUnregisterExitCallbackPlay Station Portable

Unregisters the exit callback.