pspsdk::sys

Module mem

Memory management.

Structs

GameInfo

The game information loaded to the system.

GameInfoFlags

A bitmask value with the information to what fields are set in a GameInfo.

GeLazy

Struct passed to sceKernelSetUsersystemLibWork.

HeapAllocOptions

Allocation options for the sceKernelAllocHeapMemoryWithOption.

HeapBlock

Heap block information.

HeapCreateFlag

Flags to create and configure heap behavior on sceKernelCreateHeap.

HeapId

The heap UID.

HeapInfo

The information of the current state of a heap object in the system.

LowHeapBlock

Low Heap block information.

LowHeapInfo

The information of the current state of a low heap.

LowheapInfoBlock

MemoryBlockAllocOptions

Extra options to pass to sceKernelAllocMemoryBlock.

MemoryBlockId

The memory block UID.

MemoryPartitionBasicInfo

Basic information of a memory partition.

MemoryPartitionInfo

Full information of a memory partition.

MemoryPartitionTable

Snapshot of the system memory partition layout.

UidControlBlock

Control block for UID

UidList

The UID list representation.

UidLookupFunction

Lookup table of UidControlBlock.

Enums

MemoryBlockKind

Type the specifies the kind of allocation is used for memory blocks.

MemoryPartitionId

The RAM partition ID

PspHardwareModel

PSP Hardware models.

Functions

sceKernelAllocHeapMemory⚠Play Station Portable

Allocates a memory block from a heap.

sceKernelAllocHeapMemoryWithOption⚠Play Station Portable

Allocates a memory block from a heap configured with options.

sceKernelAllocMemoryBlock⚠Play Station Portable

Allocates a memory block from the main user partition.

sceKernelAllocPartitionMemory⚠Play Station Portable

Allocates a memory block from a memory partition.

sceKernelCopyGameInfo⚠Play Station Portable

Copy the global game information to a given local structure.

sceKernelCreateHeap⚠Play Station Portable

Creates a heap.

sceKernelDeleteHeap⚠Play Station Portable

Deletes a heap associated to a heap UID.

sceKernelDeleteUID⚠Play Station Portable

Deletes a UID.

sceKernelDevkitVersionPlay Station Portable

Gets the firmware version.

sceKernelFillFreeBlockPlay Station Portable

Fills the free blocks with a value.

sceKernelFreeHeapMemory⚠Play Station Portable

Deallocates a memory block allocated from a heap.

sceKernelFreeMemoryBlock⚠Play Station Portable

Deallocates a memory block allocated with sceKernelAllocMemoryBlock.

sceKernelFreePartitionMemory⚠Play Station Portable

Deallocates a memory block allocated with sceKernelAllocPartitionMemory.

sceKernelGetAllowReplaceUmdPlay Station Portable

Gets if replacing UMD is allowed from the global GameInfo.

sceKernelGetBlockHeadAddrPlay Station Portable

Gets the address of a memory block.

sceKernelGetCompiledSdkVersionPlay Station Portable

Gets the SDK version set with sceKernelSetCompiledSdkVersion.

sceKernelGetDNASPlay Station Portable

Gets the DNAS information from the globel GameInfo.

sceKernelGetGameInfoPlay Station Portable

Gets the information of the game.

sceKernelGetId⚠Play Station Portable

Call the registered get ID function with sceKernelRegisterGetIdFunc, call it, and then unregisters it.

sceKernelGetInitialRandomValuePlay Station Portable

Gets the initial system random value.

sceKernelGetMemoryBlockAddr⚠Play Station Portable

Gets the address of a memory block.

sceKernelGetModelPlay Station Portable

Gets the hardware model of PSP.

sceKernelGetQTGP2Play Station Portable

Gets the QTGP2 value from the global GameInfo.

sceKernelGetQTGP3Play Station Portable

Gets the QTGP3 value from the global GameInfo.

sceKernelGetSystemStatusPlay Station Portable

Gets the current status of the system.

sceKernelGetUIDcontrolBlock⚠Play Station Portable

Get a UID control block.

sceKernelGetUIDcontrolBlockWithType⚠Play Station Portable

Gets a UID control block on a particular kind.

sceKernelGetUidmanCBPlay Station Portable

Gets the root of the UID list.

sceKernelHeapTotalFreeSizePlay Station Portable

Gets the total amount of free memory of a heap.

sceKernelMaxFreeMemSizePlay Station Portable

Gets the size of the largest free memory block.

sceKernelPartitionMaxFreeMemSizePlay Station Portable

Gets the size of the largest free memory block of a memory partition.

sceKernelPartitionTotalFreeMemSizePlay Station Portable

Gets the total amount of free memory of a memory partition.

sceKernelPrintf⚠Play Station Portable

Kernel printf function.

sceKernelQueryHeapInfoPlay Station Portable

Gets the information of a heap object.

sceKernelQueryLowheapInfo⚠Play Station Portable

Get the low heap information

sceKernelQueryMemoryInfoPlay Station Portable

Obtains the memory information of a given address.

sceKernelQueryMemoryPartitionInfoPlay Station Portable

Requests the information of a partition.

sceKernelRebootKernel⚠Play Station Portable

Calls the current set reboot kernel function.

sceKernelRegisterGetIdFunc⚠Play Station Portable

Register a get ID function, that will be used on sceKernelGetId.

sceKernelSetAllowReplaceUmdPlay Station Portable

Sets if replacing UMD is allowed on the global GameInfo.

sceKernelSetCompiledSdkVersionPlay Station Portable

Sets the version of the SDK with which the caller was compiled.

sceKernelSetDNASPlay Station Portable

Sets the DNAS information on the global GameInfo.

sceKernelSetDdrMemoryProtection⚠Play Station Portable

Sets the protection of a block of DDR memory.

sceKernelSetParamSfo⚠Play Station Portable

Sets the UMD Param SFO on the global GameInfo.

sceKernelSetQTGP2⚠Play Station Portable

Sets the QTGP2 value on the global GameInfo.

sceKernelSetQTGP3⚠Play Station Portable

Sets the QTGP3 value on the global GameInfo.

sceKernelSetRebootKernel⚠Play Station Portable

Sets the reboot kernel function.

sceKernelSetSystemStatus⚠Play Station Portable

Sets the status of the system.

sceKernelSetUsersystemLibWork⚠Play Station Portable

sceKernelSysMemDumpPlay Station Portable

Makes the kernel to dump the internal memory table to kernel stdout.

sceKernelSysMemDumpBlockPlay Station Portable

Makes the kernel to dump the list of memory blocks to kernel stdout.

sceKernelSysMemDumpTailPlay Station Portable

Makes the kernel to dump the tail blocks to kernel stdout.

sceKernelTotalFreeMemSizePlay Station Portable

Gets the total amount of free memory.

sceKernelTotalMemSizePlay Station Portable

Gets the total size of the main user memory partition.

Type Aliases

GeListUpdateStallAddrLazy

Function type for sceKernelSetUsersystemLibWork.

Unions

UidControlBlockNext

Inner union of UidControlBlock.