pspsdk::sys

Module module

Module management.

Structs

KernelModuleInfo

Holds information about a module.

KernelModuleInfoV1

Holds information about a module compatible with firmware version less or equal to v1.50.

LoadModuleOptions

Options to configure the module loading behavior.

ModuleAttributes

The attributes for modules.

ModuleId

The module UID.

StartModuleOptions

Options to configure the module loading behavior.

Enums

LoadModulePosition

Options of the module placement in memory on load time.

Functions

sceKernelGetModuleIdPlay Station Portable

Gets the module ID of the caller module.

sceKernelGetModuleIdByAddressPlay Station Portable

Gets the module ID of the module occupying the given address.

sceKernelGetModuleIdList⚠Play Station Portable

Get a list of module UIDs.

sceKernelLoadModule⚠Play Station Portable

Loads a module.

sceKernelLoadModuleBufferBootInitBtcnf⚠Play Station Portable

Load module from a buffer with the Boot Init BTCNF apitype.

sceKernelLoadModuleBufferUsbWlan⚠Play Station Portable

Load a module from a buffer using the USB/WLAN API.

sceKernelLoadModuleByIDPlay Station Portable

Loads a module from a file ID.

sceKernelLoadModuleByIDWithBlockOffsetPlay Station Portable

Loads a module at a offset in a memory block from a file ID.

sceKernelLoadModuleMs⚠Play Station Portable

Loads a module from the Memory Stick.

sceKernelLoadModuleWithBlockOffset⚠Play Station Portable

Loads a module at a offset in a memory block.

sceKernelQueryModuleInfoPlay Station Portable

Queries the information about module from its UID.

sceKernelSelfStopUnloadModule⚠Play Station Portable

Stops and unloads the current module.

sceKernelStartModule⚠Play Station Portable

Starts as loaded module.

sceKernelStopModule⚠Play Station Portable

Stops a running module.

sceKernelStopUnloadSelfModule⚠Play Station Portable

Stops and unloads the current module.

sceKernelStopUnloadSelfModuleWithStatus⚠

Stops and unloads the current module with a exit code.

sceKernelUnloadModulePlay Station Portable

Unloads a stopped module.