pspsdk::sys

Module power

Power management.

Structs

PowerCallbackArg

Bitflags that are passed to the arg parameter of CallbackFunction when registered with scePowerRegisterCallback.

PowerCallbackSlot

Valid power callback slot.

Enums

ExclusiveWlanMode

Represents the possible exclusive WLAN modes.

PowerLockKind

Possible power lock values.

PowerTick

Possible power tick configurations.

WlanCoexistenceClock

Represents the device support of the maximum PLL clock frequency while WLAN is in use.

Functions

scePowerCheckWlanCoexistenceClockPlay Station Portable

Gets what is the maximum frequency the PLL while WLAN is in use.

scePowerCheckWlanConditionPlay Station Portable

Checks if the WLAN can be activated.

scePowerGetBacklightMaximumPlay Station Portable

Gets the current maximum backlight level.

scePowerGetBatteryChargingStatusPlay Station Portable

Gets the status of the battery charging

scePowerGetBatteryElecPlay Station Portable

Gets a unknown battery information.

scePowerGetBatteryFullCapacityPlay Station Portable

Gets the battery full capacity in milliampere hour (mAh).

scePowerGetBatteryLifePercentPlay Station Portable

Gets the remaining battery life in percent.

scePowerGetBatteryLifeTimePlay Station Portable

Gets the remaining battery life in minutes.

scePowerGetBatteryRemainCapacityPlay Station Portable

Gets the battery remaining capacity in milliampere hour (mAh).

scePowerGetBatteryTempPlay Station Portable

Gets the current temperature of the battery in degree celsius (°C).

scePowerGetBatteryVoltPlay Station Portable

Gets the battery voltage level.

scePowerGetBusClockFrequencyFloatPlay Station Portable

Gets the current BUS clock in MHz.

scePowerGetBusClockFrequencyIntPlay Station Portable

Gets the current BUS clock in MHz.

scePowerGetCpuClockFrequencyFloatPlay Station Portable

Gets the current CPU clock in MHz.

scePowerGetCpuClockFrequencyIntPlay Station Portable

Gets the current CPU clock in MHz.

scePowerGetIdleTimerPlay Station Portable

Gets the idle timer.

scePowerGetPllClockFrequencyFloatPlay Station Portable

Gets the current PLL clock in MHz.

scePowerGetPllClockFrequencyIntPlay Station Portable

Gets the current PLL clock in MHz.

scePowerIdleTimerDisablePlay Station Portable

Disables the idle timer.

scePowerIdleTimerEnablePlay Station Portable

Enables the idle timer.

scePowerIsBatteryChargingPlay Station Portable

Checks if the battery is charging.

scePowerIsBatteryExistPlay Station Portable

Checks if a battery is present.

scePowerIsLowBatteryPlay Station Portable

Checks if the battery charge level is low.

scePowerIsPowerOnlinePlay Station Portable

Checks if power is supplied from external power source.

scePowerIsRequestPlay Station Portable

Checks if a power request is uncompleted.

scePowerIsSuspendRequiredPlay Station Portable

Checks if a suspend is required.

scePowerLockPlay Station Portable

Locks power state of the device.

scePowerRegisterCallbackPlay Station Portable

Register Power callback function.

scePowerRequestStandbyPlay Station Portable

Request the system to go into standby mode.

scePowerRequestSuspendPlay Station Portable

Request the system to go into suspend mode.

scePowerSetBusClockFrequencyPlay Station Portable

Sets the BUS clock.

scePowerSetClockFrequencyPlay Station Portable

Sets the clock frequencies.

scePowerSetCpuClockFrequencyPlay Station Portable

Sets the CPU clock.

scePowerSetExclusiveWlanPlay Station Portable

Sets the exclusive WLAN mode before activating WLAN.

scePowerTickPlay Station Portable

Configures the system to cancel the count of idle timers to prevent the system to enter in power save state (entirely or partial).

scePowerUnlockPlay Station Portable

Unlocks power state of the device.

scePowerUnregisterCallbackPlay Station Portable

Unregisters Power callback function.

scePowerVolatileMemLock⚠Play Station Portable

Locks and grants access to the device volatile memory (blocking).

scePowerVolatileMemTryLock⚠Play Station Portable

Locks and grants access to the device volatile memory (non-blocking).

scePowerVolatileMemUnlock⚠Play Station Portable

Unlock access to the device volatile memory.

scePowerWaitRequestCompletionPlay Station Portable

Waits for the completion of the previous request.

scePowerWlanActivatePlay Station Portable

Activates the WLAN device.

scePowerWlanDeactivatePlay Station Portable

Deactivates the WLAN device.