pspsdk::sys

Struct SceResult

pub struct SceResult<T>(/* private fields */);

A type that represents the return value of many PSP OS APIs.

If its value is in the range of SceError ( 0x80000001..=0xFFFFFFFF), then it is an error result, and a success value otherwise.

Cloning behavior

Cloning this type is cheap, the reason it is not [Copy] is to incentivize the handling of the error values.

Implementations

impl<T> SceResult<T>

pub const fn new(raw: u32) -> Self

Create a new SceResult.

pub const fn is_ok(&self) -> bool

Returns true if the result is a Ok value.

pub const fn is_err(&self) -> bool

Returns true if the result is a error value.

impl<T: SceResultOk> SceResult<T>

pub fn into_result(self) -> Result<T, SceError>

Turn the SceResult into a [Result] type.

pub fn ok(self) -> Option<T>

Converts from SceResult to [Option<T>].

Converts self into an [Option<T>], consuming self, and discarding the error, if any.

pub fn err(self) -> Option<SceError>

Converts from SceResult<T> to [Option<SceError>].

Converts self into an [Option<SceError>], consuming self, and discarding the success value, if any.

pub fn map<U, F>(self, op: F) -> Result<U, SceError>
where F: FnOnce(T) -> U,

contained [Ok] value, leaving an [Err] value untouched.

This function can be used to compose the results of two functions.

pub fn map_or<U, F>(self, default: U, f: F) -> U
where F: FnOnce(T) -> U,

Returns the provided default (if Err), or applies a function to the contained value (if Ok).

Arguments passed to map_or are eagerly evaluated; if you are passing the result of a function call, it is recommended to use map_or_else, which is lazily evaluated.

pub fn map_or_else<U, D, F>(self, default: D, f: F) -> U
where D: FnOnce(SceError) -> U, F: FnOnce(T) -> U,

Maps a SceResult<T> to U by applying fallback function default to a contained Err value, or function f to a contained Ok value.

This function can be used to unpack a successful result while handling an error.

pub fn map_or_default<U, F>(self, f: F) -> U
where F: FnOnce(T) -> U, U: Default,

Maps a SceResult<T> to a U by applying function f to the contained value if the result is Ok], otherwise if Err, returns the default value for the type U.

pub fn map_err<F, O>(self, op: O) -> Result<T, F>
where O: FnOnce(SceError) -> F,

Maps a ScwResult<T> to Result<T, F> by applying a function to a contained Err value, leaving an Ok value untouched.

This function can be used to pass through a successful result while handling an error.

pub fn inspect<F>(self, f: F) -> Self
where F: FnOnce(&T),

Calls a function with a reference to the contained value if Ok value.

Returns the original result.

pub fn inspect_err<F>(self, f: F) -> Self
where F: FnOnce(SceError),

Calls a function with a reference to the contained value if Err value.

Returns the original result.

impl SceResult<()>

pub const OK: Self

Everything ok.