pspsdk::sys::suspend

Function sceKernelVolatileMemTryLock

pub unsafe extern "C" fn sceKernelVolatileMemTryLock(
    unk: u32,
    ptr: *mut *mut u8,
    size: &mut SceSize,
) -> SceResult<()>

Available on Play Station Portable only.

Locks and grants access to the device volatile memory (non-blocking).

If the volatile memory is currently being used by other processes, this function will return a error.

Parameters

unk: Unknown. Always zero on reversed code.

ptr: A pointer to receive the head pointer of the volatile memory.

size: A reference to receive the size of the volatile memory.

Return Value

Ok value on success, error value otherwise.

Specifically SceError::POWER_CANNOT_LOCK_VMEM when volatile memory is in use by other processes.

Precautions

The volatile memory is used by the PSP system for the utility library and to save eDRAM during suspend/resume time and those will be blocked while this memory is being used by your software. For that reason, it is best to unlock (with sceKernelVolatileMemUnlock) as soon as you done with using that piece of memory.

If you need to use this memory for long periods and want to maintain such functionalities working, you, theoretically, can create and register a power callback that handles values of the arg parameter of CallbackFunction (namely PowerCallbackArg::Standby and PowerCallbackArg::Suspending to unlock and PowerCallbackArg::ResumeComplete to lock again).

Firmware Version

This API was introduced on PSP firmware version 1.50.