pspsdk::sys::sync

Struct LwMutex

pub struct LwMutex { /* private fields */ }

Available on crate feature non-stub-code only.

A raw lightweight mutex based on sys::thread lightweight mutex API.

This implementation has the advantage to be faster and more lightweight process-wise, but it has the downside of the entire mutex structure requiring to live in user RAM partition (this is automatically handled by this type), i.e. it uses slightly more user RAM space.

Another niche downside of this type is that the required API was only introduced on PSP firmware version 3.95, so if the software was made to run on lower firmware can’t use this kind of mutex. An alternative is to use either Mutex (available on firmware >= 2.70) or SemaMutex (always available, i.e. since 1.00).

Implementations

impl LwMutex

pub const fn new() -> Self

impl LwMutex

pub fn lock(&self)

pub fn try_lock(&self) -> bool

pub unsafe fn unlock(&self)