pspsdk::sys::sync

Struct LwRwLock

pub struct LwRwLock { /* private fields */ }

Available on crate feature non-stub-code only.

A raw lightweight rwlock with writer-preference based on sys::thread lightweight mutex API.

This implementation has the advantage to be faster and more lightweight process-wise, but it has the downside of the entire rwlock structure requiring to live in user RAM partition (this is automatically handled by this type), i.e. it uses slightly more user RAM space.

Another niche downside of this type is that the required API was only introduced on PSP firmware version 3.95, so if the software was made to run on lower firmware can’t use this kind of mutex. An alternative is to use either RwLock (available on firmware >= 2.70) or SemaRwLock (always available, i.e. since 1.00).

Implementations

impl LwRwLock

pub const fn new() -> Self