pspsdk::sys

Module thread

Thread operations and management.

Re-exports

pub use crate::sys::usersystemlib::sceKernelGetTlsAddr;

pub use crate::sys::usersystemlib::sceKernelLockLwMutex;

pub use crate::sys::usersystemlib::sceKernelLockLwMutexCB;

pub use crate::sys::usersystemlib::sceKernelReferLwMutexStatus;

pub use crate::sys::usersystemlib::sceKernelTryLockLwMutex;

pub use crate::sys::usersystemlib::sceKernelUnlockLwMutex;

Structs

AlarmId

The alarm UID.

AlarmInfo

The information of the current state of a alarm timer.

CallbackId

The callback UID, created with sceKernelCreateCallback.

CallbackInfo

The information of the current state of a callback.

EventFlagAttributes

Attributes for event flag creation.

EventFlagId

The event flag UID, created with sceKernelCreateEventFlag.

EventFlagInfo

The information of the current state of a event flag.

EventFlagOptions

Event Flag options.

EventFlagWaitKinds

The event flag wait kinds.

FplAttributes

Attributes for fixed-sized memory pool creation.

FplId

The fixed-sized memory pool UID.

FplInfo

The information of the current state of a fixed-sized memory pool.

FplOptions

Fixed-sized memory pool options.

KtlsId

The kernel TLS UID.

LwMutexId

The lightweight mutex UID.

LwMutexInfo

The information of the current state of a lightweight mutex.

LwMutexOptions

Lightweight mutex options.

LwMutexWorkArea

The work area for the lightweight mutex to work on.

MsgBoxAttributes

Attributes for message box creation.

MsgBoxId

The message box UID.

MsgBoxInfo

The information of the current state of a message box.

MsgBoxOptions

Message box options.

MsgPacket

The header of the a message package with user data.

MsgPipeAttributes

Attributes for message pipe creation.

MsgPipeId

The message pipe UID.

MsgPipeInfo

The information of the current state of a message pipe.

MsgPipeOptions

Message pipe options.

MutexAttributes

Attributes for mutex creation.

MutexId

The mutex UID, created with sceKernelCreateMutex.

MutexInfo

The information of the current state of a mutex.

MutexOptions

Mutex options.

SemaId

The semaphore UID, created with sceKernelCreateSema.

SemaphoreAttributes

Attributes for semaphore creation.

SemaphoreInfo

The information of the current state of a semaphore.

SemaphoreOptions

Semaphore options.

SystemInfo

The information of the current state of the system.

SystemStatus

ThreadAttributes

Attributes for threads.

ThreadEventId

The thread event UID.

ThreadEventInfo

The information of the current state of a thread event.

ThreadEvents

The thread events that an thread event handler may be called.

ThreadId

The thread UID, created with sceKernelCreateThread.

ThreadInfo

The status information for a thread.

ThreadOptions

Extra options used when creating threads.

ThreadRunStatus

Statistics about a running thread.

ThreadState

The possible states/status that a PSP thread can be.

TlsPoolAttributes

Attributes for TLS memory pool creation.

TlsPoolId

The user TLS pool UID.

TlsPoolInfo

The information of the current state of a TLS memory pool.

TlsPoolOptions

TLS memory pool options.

VirtualTimerId

The virtual timer UID.

VirtualTimerInfo

The information of the current state of a alarm timer.

VirtualTimerOptions

Virtual Timer options.

VplAttributes

Attributes for variable-sized memory pool creation.

VplId

The variable-sized memory pool UID.

VplInfo

The information of the current state of a variable-sized memory pool.

VplOptions

Variable-sized memory pool options.

Enums

CallbackCheckStatus

The possible return values of sceKernelCheckCallback.

CallbackTermState

The termination state of a CallbackFunction.

MsgPipeWaitKind

Possible wait strategies for message pipe.

ThreadEventHandlerTermState

The termination state of a ThreadEventHandler.

ThreadIdKind

The kinds of possible thread ID

ThreadWaitKind

The possible kinds of wait state of the thread.

VirtualTimerState

The possible running states of a virtual timer.

Functions

_sceKernelAllocateTlspl⚠Play Station Portable

Allocates the TLS memory pool.

_sceKernelLockLwMutexPlay Station Portable

Locks a lightweight mutex a number of times.

_sceKernelLockLwMutexCBPlay Station Portable

Locks a lightweight mutex a number of times, but service any callbacks as necessary.

_sceKernelTryLockLwMutexPlay Station Portable

Tries to lock a lightweight mutex a number of times.

_sceKernelUnlockLwMutexPlay Station Portable

Unlocks a mutex a number of times.

sceKernelAllocateFpl⚠Play Station Portable

Allocates a memory block from a fixed-sized memory pool.

sceKernelAllocateFplCB⚠Play Station Portable

Allocates a memory block from a fixed-sized memory pool, but service any callbacks as necessary.

sceKernelAllocateKTLS⚠Play Station Portable

Setup the KTLS allocator.

sceKernelAllocateVpl⚠Play Station Portable

Allocates a memory block from a variable-sized memory pool.

sceKernelAllocateVplCB⚠Play Station Portable

Allocates a memory block from a variable-sized memory pool, but service any callbacks as necessary.

sceKernelCancelAlarmPlay Station Portable

Cancels an alarm.

sceKernelCancelCallbackPlay Station Portable

Cancels all notifications that were reported to a callback.

sceKernelCancelEventFlagPlay Station Portable

Cancels the wait of a event flag.

sceKernelCancelFplPlay Station Portable

Cancels the wait of a fixed-sized memory pool.

sceKernelCancelMsgPipePlay Station Portable

Cancels the wait of message pipe.

sceKernelCancelMutexPlay Station Portable

Cancels the wait state of threads waiting on a mutex.

sceKernelCancelReceiveMbxPlay Station Portable

Cancels the wait of a message box.

sceKernelCancelSemaPlay Station Portable

Cancels the wait of a semaphore.

sceKernelCancelVTimerHandlerPlay Station Portable

Cancels the handler set for a virtual timer.

sceKernelCancelVplPlay Station Portable

Cancels the wait of a variable-sized memory pool.

sceKernelCancelWakeupThreadPlay Station Portable

Cancels all the wake-up requests from a thread.

sceKernelChangeCurrentThreadAttrPlay Station Portable

Changes the thread attributes of the calling thread.

sceKernelChangeThreadPriorityPlay Station Portable

Changes the thread current priority of a thread.

sceKernelCheckCallbackPlay Station Portable

Checks if a callback by the calling thread has been notified.

sceKernelCheckThreadKernelStackPlay Station Portable

Checks the calling kernel thread stack.

sceKernelCheckThreadStackPlay Station Portable

Gets the remaining free size of the calling thread stack (?)

sceKernelClearEventFlagPlay Station Portable

Clears a bit pattern of a event flag.

sceKernelCreateCallback⚠Play Station Portable

Creates a new callback.

sceKernelCreateEventFlag⚠Play Station Portable

Creates a event flag.

sceKernelCreateFpl⚠Play Station Portable

Creates a new fixed-sized memory pool.

sceKernelCreateLwMutex⚠

Creates a new lightweight mutex.

sceKernelCreateMbx⚠Play Station Portable

Creates a new message box.

sceKernelCreateMsgPipe⚠Play Station Portable

Creates a new message pipe.

sceKernelCreateMutex⚠Play Station Portable

Creates a new mutex.

sceKernelCreateSema⚠Play Station Portable

Creates a new semaphore

sceKernelCreateThread⚠Play Station Portable

Create a thread.

sceKernelCreateTlspl⚠

Creates a new user TLS memory pool.

sceKernelCreateVTimer⚠Play Station Portable

Creates a new virtual timer.

sceKernelCreateVpl⚠Play Station Portable

Creates a new variable-sized memory pool.

sceKernelDelaySysClockThreadPlay Station Portable

Delays the calling thread by a specified number of SystemClocks.

sceKernelDelaySysClockThreadCBPlay Station Portable

Delays the calling thread by a specified number of SystemClocks, but service any callbacks as necessary.

sceKernelDelayThreadPlay Station Portable

Delays the calling thread by a specified number of microseconds.

sceKernelDelayThreadCBPlay Station Portable

Delays the calling thread by a specified number of microseconds, but service any callbacks as necessary.

sceKernelDeleteCallbackPlay Station Portable

Deletes a callback.

sceKernelDeleteEventFlagPlay Station Portable

Deletes a event flag.

sceKernelDeleteFplPlay Station Portable

Deletes a fixed-sized memory pool.

sceKernelDeleteLwMutexPlay Station Portable

Deletes a lightweight mutex.

sceKernelDeleteMbx⚠Play Station Portable

Deletes a message box.

sceKernelDeleteMsgPipePlay Station Portable

Deletes a message pipe.

sceKernelDeleteMutexPlay Station Portable

Deletes a mutex.

sceKernelDeleteSemaPlay Station Portable

Deletes a semaphore.

sceKernelDeleteThread⚠Play Station Portable

Deletes a thread.

sceKernelDeleteTlsplPlay Station Portable

Deletes a user TLS memory pool.

sceKernelDeleteVTimerPlay Station Portable

Deletes a virtual timer.

sceKernelDeleteVplPlay Station Portable

Deletes a variable-sized memory pool.

sceKernelDonateWakeupThreadPlay Station Portable

Transfers the wake-up request count of the calling thread to another thread.

sceKernelExitDeleteThreadPlay Station Portable

Exits and deletes a thread.

sceKernelExitThreadPlay Station Portable

Exits a thread.

sceKernelExtendKernelStackPlay Station Portable

Temporarily extends the kernel thread stack an executes a function with extended stack size.

sceKernelExtendThreadStackPlay Station Portable

Temporarily extends the thread stack an executes a function with extended stack size.

sceKernelFreeFpl⚠Play Station Portable

Deallocates a memory block from a fixed-sized memory pool.

sceKernelFreeKTLSPlay Station Portable

Deallocates a KTLS allocator.

sceKernelFreeVpl⚠Play Station Portable

Deallocates a memory block from a variable-sized memory pool.

sceKernelGetCallbackCountPlay Station Portable

Gets the notification count of a callback.

sceKernelGetKTLSPlay Station Portable

Gets the KTLS of the calling thread.

sceKernelGetSystemStatusFlagPlay Station Portable

Gets the system status flag.

sceKernelGetSystemTimePlay Station Portable

Gets the system time.

sceKernelGetSystemTimeLowPlay Station Portable

Gets the low part of the system time.

sceKernelGetSystemTimeWidePlay Station Portable

Gets the system time as raw wide integer.

sceKernelGetThreadCurrentPriorityPlay Station Portable

Gets the current priority of the calling thread.

sceKernelGetThreadExitStatusPlay Station Portable

Gets the exit status of a thread.

sceKernelGetThreadIdPlay Station Portable

Gets the thread UID of the calling thread.

sceKernelGetThreadKTLSPlay Station Portable

Gets the KTLS of a thread.

sceKernelGetThreadKernelStackFreeSizePlay Station Portable

Gets the available stack space on a kernel thread.

sceKernelGetThreadStackFreeSizePlay Station Portable

Gets the unused thread stack size of a thread.

sceKernelGetThreadmanIdList⚠Play Station Portable

Gets a list of UIDs from threadman module.

sceKernelGetThreadmanIdTypePlay Station Portable

Gets the threadman UID kind of an ID.

sceKernelGetUserLevelPlay Station Portable

Gets the user level of the calling thread.

sceKernelGetVTimerBasePlay Station Portable

Gets the base time of a virtual timer.

sceKernelGetVTimerBaseWidePlay Station Portable

Gets the base time of a virtual timer.

sceKernelGetVTimerTimePlay Station Portable

Gets the current time of a virtual timer.

sceKernelGetVTimerTimeWidePlay Station Portable

Gets the current time of a virtual timer.

sceKernelIsUserModeThreadPlay Station Portable

Gets if the calling thread is a user mode thread.

sceKernelLockMutexPlay Station Portable

Locks a mutex a number of times.

sceKernelLockMutexCBPlay Station Portable

Locks a mutex a number of times, but service any callbacks as necessary.

sceKernelNotifyCallbackPlay Station Portable

Notifies a callback.

sceKernelPollEventFlagPlay Station Portable

Polls for a bit pattern of a event flag.

sceKernelPollMbx⚠Play Station Portable

Polls if the message has arrived in a message box.

sceKernelPollSemaPlay Station Portable

Polls a semaphore if the target value has reached.

sceKernelReceiveMbx⚠Play Station Portable

Waits to receive a message to a message box.

sceKernelReceiveMbxCB⚠Play Station Portable

Waits to receive a message to a message box, but service any callbacks as necessary.

sceKernelReceiveMsgPipe⚠Play Station Portable

Waits to receive a message from a message pipe.

sceKernelReceiveMsgPipeCB⚠Play Station Portable

Waits to receive a message from a message pipe, but service any callbacks as necessary.

sceKernelReferAlarmStatusPlay Station Portable

Gets the current state of a alarm.

sceKernelReferCallbackStatusPlay Station Portable

Gets the current state of a callback.

sceKernelReferEventFlagStatusPlay Station Portable

Gets the current state of a event flag.

sceKernelReferFplStatusPlay Station Portable

Gets the current state of a fixed-sized memory pool.

sceKernelReferLwMutexStatusByIDPlay Station Portable

Gets the current state of a lightweight mutex by its UID.

sceKernelReferMbxStatusPlay Station Portable

Gets the current state of a message box.

sceKernelReferMsgPipeStatusPlay Station Portable

Gets the current state of a message pipe.

sceKernelReferMutexStatusPlay Station Portable

Gets the current state of a mutex.

sceKernelReferSemaStatusPlay Station Portable

Gets the current state of a semaphore.

sceKernelReferSystemStatusPlay Station Portable

Gets the current state of the system.

sceKernelReferThreadEventHandlerStatusPlay Station Portable

Gets the current state of a thread event handler.

sceKernelReferThreadRunStatusPlay Station Portable

Gets the current runtime status of a thread.

sceKernelReferThreadStatusPlay Station Portable

Gets the status information of a thread.

sceKernelReferTlsplStatusPlay Station Portable

Gets the current state of a user TLS memory pool.

sceKernelReferVTimerStatusPlay Station Portable

Gets the current state of a virtual timer.

sceKernelReferVplStatusPlay Station Portable

Gets the current state of a variable-sized memory pool.

sceKernelRegisterThreadEventHandler⚠Play Station Portable

Registers a thread event handler.

sceKernelReleaseThreadEventHandlerPlay Station Portable

Releases a thread event handler.

sceKernelReleaseWaitThreadPlay Station Portable

Forces a thread to get out of a Wait state.

sceKernelResumeDispatchThreadPlay Station Portable

Resumes the dispatch thread.

sceKernelResumeThreadPlay Station Portable

Resumes a thread that is in a Suspend state.

sceKernelRotateThreadReadyQueuePlay Station Portable

Rotate thread ready queue at a given priority.

sceKernelSendMbx⚠Play Station Portable

Sends a message to a message box.

sceKernelSendMsgPipe⚠Play Station Portable

Sends a message to a message pipe.

sceKernelSendMsgPipeCB⚠Play Station Portable

Sends a message to a message pipe, but service any callbacks as necessary.

sceKernelSetAlarm⚠Play Station Portable

Creates an alarm.

sceKernelSetEventFlagPlay Station Portable

Sets a bit pattern on a even flag.

sceKernelSetSysClockAlarm⚠Play Station Portable

Creates an alarm using SystemClock.

sceKernelSetVTimerHandler⚠Play Station Portable

Sets the virtual time handler and the schedule to execute it.

sceKernelSetVTimerHandlerWide⚠Play Station Portable

Sets the virtual time handler and the schedule to execute it.

sceKernelSetVTimerTimePlay Station Portable

Sets the current time of a virtual timer.

sceKernelSetVTimerTimeWidePlay Station Portable

Sets the current time of a virtual timer.

sceKernelSignalSemaPlay Station Portable

Sends a signal to a semaphore.

sceKernelSleepThreadPlay Station Portable

Makes the calling thread to enter in a Wait state.

sceKernelSleepThreadCBPlay Station Portable

Makes the calling thread to enter in a Wait state but service any callbacks as necessary.

sceKernelStartThread⚠Play Station Portable

Starts the execution of a created thread.

sceKernelStartVTimerPlay Station Portable

Starts a virtual timer.

sceKernelStopVTimerPlay Station Portable

Stops a virtual timer.

sceKernelSuspendAllUserThreadsPlay Station Portable

Puts all user mode threads in the system in a Suspend state.

sceKernelSuspendDispatchThreadPlay Station Portable

Suspends the dispatch thread.

sceKernelSuspendThreadPlay Station Portable

Puts a thread in a Suspend state.

sceKernelSysClock2USecPlay Station Portable

Converts time in SystemClock to seconds and microseconds.

sceKernelSysClock2USecWidePlay Station Portable

Converts raw system time to seconds and microseconds.

sceKernelTerminateDeleteThread⚠Play Station Portable

Forcibly terminates and deletes a thread.

sceKernelTerminateThread⚠Play Station Portable

Forcibly terminates a thread.

sceKernelTryAllocateFpl⚠Play Station Portable

Tries to allocates a memory block from a fixed-sized memory pool.

sceKernelTryAllocateVpl⚠Play Station Portable

Tries to allocates a memory block from a variable-sized memory pool.

sceKernelTryLockMutexPlay Station Portable

Tries to lock a mutex a number of times.

sceKernelTryReceiveMsgPipe⚠Play Station Portable

Tries to receive a message from a message pipe.

sceKernelTrySendMsgPipe⚠Play Station Portable

Tries to send a message to a message pipe.

sceKernelUSec2SysClockPlay Station Portable

Converts time in microseconds to SystemClock.

sceKernelUSec2SysClockWidePlay Station Portable

Converts time in microseconds to raw system clock.

sceKernelUnlockMutexPlay Station Portable

Unlocks a mutex a number of times.

sceKernelWaitEventFlagPlay Station Portable

Waits for a bit pattern of a event flag.

sceKernelWaitEventFlagCBPlay Station Portable

Waits for a bit pattern of a event flag, but service any callbacks as necessary.

sceKernelWaitSemaPlay Station Portable

Locks a semaphore until a target value or a timeout is reached.

sceKernelWaitSemaCBPlay Station Portable

Locks a semaphore until a target value or a timeout is reached, but service any callbacks as necessary.

sceKernelWaitThreadEndPlay Station Portable

Put the calling thread in a Wait state until a thread if finished or the timeout is reached.

sceKernelWaitThreadEndCBPlay Station Portable

Put the calling thread in a Wait state until a thread if finished or the timeout is reached, but service any callbacks as necessary.

sceKernelWakeupThreadPlay Station Portable

Wakes a thread that was previously put in the Wait state.

Type Aliases

AlarmHandler

The alarm entry function.

CallbackFunction

The callback entry function.

ExtendStackFunc

Function of a ExtendStack functions

KtlsAllocFunc

The KTLS allocation function.

ThreadEntryFn

The thread entry function.

ThreadEventHandler

The thread event entry function.

VirtualTimerHandler

The virtual timer entry function.

VirtualTimerHandlerWide

The virtual timer entry function.