pspsdk::sys::thread

Function sceKernelRegisterThreadEventHandler

pub unsafe extern "C" fn sceKernelRegisterThreadEventHandler(
    name: *const u8,
    thread_id: ThreadId,
    event_mask: ThreadEvents,
    handler: ThreadEventHandler,
    common: *mut c_void,
) -> SceResult<ThreadEventId>

Available on Play Station Portable only.

Registers a thread event handler.

Parameters

name [[In parameter]]: The name assigned to the new thread event. Only used for debug.

thread_id: The thread UID of the thread which the handler should be called. The ThreadId::CALLING can be used to specify the calling thread UID and the ThreadId::ALL_USER can be used to set to be called for all user threads.

event_mask: The thread events which the handler should be called.

handler [[In parameter]]: A function pointer to the entry function for the handler.

common [[InOut parameter]]: A pointer to memory shared with the thread event handler.

Return Value

Returns the thread event UID on success, error value otherwise.