pspsdk::sys

Module time

Time operation and management.

Structs

DateTime

PSP Date Time structure.

SystemClock

The 64-bit system clock type.

Enums

DayOfWeek

The possible values for day of the week.

Month

The possible month values

Functions

sceRtcCheckValidPlay Station Portable

Checks if a PSP date-time is valid.

sceRtcConvertLocalTimeToUtcPlay Station Portable

Converts a UTC time tick to local-time time tick.

sceRtcConvertUtcToLocalTimePlay Station Portable

Converts a UTC time tick to local-time time tick.

sceRtcFormatRFC2822⚠Play Station Portable

Formats the tick time in the RFC2822 format, adjusted for specified timezone.

sceRtcFormatRFC3339⚠Play Station Portable

Formats the tick time in the RFC3339 format, adjusted for specified timezone.

sceRtcFormatRFC2822LocalTime⚠Play Station Portable

Formats the tick time in the RFC2822 format in the currently set local time.

sceRtcFormatRFC3339LocalTime⚠Play Station Portable

Formats the tick time in the RFC3339 format in the currently set local time.

sceRtcGetAccumulativeTimePlay Station Portable

Gets the ticks from the last time a battery drained system reset has occurred.

sceRtcGetAlarmTickPlay Station Portable

Gets the current RTC alarm tick count.

sceRtcGetCurrentClock⚠Play Station Portable

Gets the current clock, adjusted for specified timezone.

sceRtcGetCurrentClockLocalTime⚠Play Station Portable

Gets the current clock in the currently set local time.

sceRtcGetCurrentNetworkTickPlay Station Portable

Gets the current network tick count. (It seems like it needs to be logged in PSN to not return a error).

sceRtcGetCurrentTickPlay Station Portable

Gets the current tick count.

sceRtcGetDayOfWeekPlay Station Portable

Gets the day of the week of a date.

sceRtcGetDaysInMonthPlay Station Portable

Gets the number of days of a month in a year.

sceRtcGetDosTimePlay Station Portable

Converts a date-time to DOS time.

sceRtcGetLastAdjustedTimePlay Station Portable

Gets the tick of the last time the system time was adjusted.

sceRtcGetLastReincarnatedTimePlay Station Portable

Gets the ticks from the last time a battery drained system reset has occurred based on the system time.

sceRtcGetTickPlay Station Portable

Converts the a DateTime to ticks.

sceRtcGetTickResolutionPlay Station Portable

Gets the resolution of the real-time clock tick.

sceRtcGetTime64_tPlay Station Portable

Gets the POSIX time64_t from a PSP date-time.

sceRtcGetTime_tPlay Station Portable

Gets the POSIX time_t from a PSP date-time.

sceRtcGetWin32FileTimePlay Station Portable

Gets the WIN32 FILETIME from a PSP date-time.

sceRtcIsAlarmedPlay Station Portable

Checks if the RTC alarm is alarmed.

sceRtcIsLeapYearPlay Station Portable

Checks if the given year is a leap year.

sceRtcParseDateTime⚠Play Station Portable

Parses a formatted buffer text date-time in ticks.

sceRtcParseRFC3339⚠Play Station Portable

Parses a RFC3339-formatted buffer text date-time in ticks.

sceRtcRegisterCallbackPlay Station Portable

Registers a callback to be executed when RTC alarm changes state.

sceRtcSetAlarmTickPlay Station Portable

Sets the RTC alarm tick count.

sceRtcSetDosTime⚠Play Station Portable

Converts a MS DOS time to date-time.

sceRtcSetTick⚠Play Station Portable

Converts a time in ticks to DateTime.

sceRtcSetTime64_t⚠Play Station Portable

Converts a POSIX time64_t time to a PSP date-time.

sceRtcSetTime_t⚠Play Station Portable

Converts a POSIX time_t time to a PSP date-time.

sceRtcSetWin32FileTime⚠Play Station Portable

Converts a WIN32 FILETIME time to a PSP date-time.

sceRtcTickAddDays⚠Play Station Portable

Add days to a tick.

sceRtcTickAddHours⚠Play Station Portable

Add hours to a tick.

sceRtcTickAddMicroseconds⚠Play Station Portable

Add microseconds to a tick.

sceRtcTickAddMinutes⚠Play Station Portable

Add minutes to a tick.

sceRtcTickAddMonths⚠Play Station Portable

Add months to a tick.

sceRtcTickAddSeconds⚠Play Station Portable

Add seconds to a tick.

sceRtcTickAddTicks⚠Play Station Portable

Add two ticks.

sceRtcTickAddWeeks⚠Play Station Portable

Add weeks to a tick.

sceRtcTickAddYears⚠Play Station Portable

Add years to a tick.

sceRtcUnregisterCallbackPlay Station Portable

Unregisters a RTC alarm callback previously registered.