pspsdk::sys

Trait SceIntoOkValue

pub unsafe trait SceIntoOkValue: Sized + Sealed {
    // Required method
    fn into_ok_value(self) -> u32;
}

Trait that specifies that it can safely be turned to a SceResult ok-valid value.

Safety

For this trait to be correct, the implementation must:

Be valid for the 0..=0x7FFFFFFF range

Have a max size of 4 bytes

If a struct, be repr(transparent)

Required Methods

fn into_ok_value(self) -> u32

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".