pspsdk::sys

Trait SceResultOk

pub unsafe trait SceResultOk: Sized + Sealed {
    // Required method
    unsafe fn handle_ok_value(ok_value: u32) -> Option<Self>;

    // Provided method
    unsafe fn handle_ok_value64(ok_value: u64) -> Option<Self> { ... }
}

Trait of types that can be the ok result of SceResult or SceResult64.

Safety

For this trait to be correct, the implementation must:

Be valid for the 0..=0x7FFFFFFF range

Have a max size of 4 bytes

If a struct, be repr(transparent)

Required Methods

unsafe fn handle_ok_value(ok_value: u32) -> Option<Self>

Turn the Ok value range ((0,0x7FFFFFFF]) into a option.

The ok_value is always in the ok value range when used by SceResult.

Safety

SceResult methods that use it will ensure to pass only the valid range and the implementation is free to not check for values outside of that range.

For callers outside of SceResult, it must ensure to pass a valid ok_value.

Provided Methods

unsafe fn handle_ok_value64(ok_value: u64) -> Option<Self>

Turn the Ok value range ((0,0xFFFFFFFF_7FFFFFFF]) into a option.

The ok_value is always in the ok value range when used by SceResult64.

Safety

SceResult64 methods that use it will ensure to pass only the valid range and the implementation is free to not check for values outside of that range.

For callers outside of SceResult64, it must ensure to pass a valid ok_value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".