pspsdk::sys

Module usersystemlib

The usersustemlib.prx exported functions.

Functions

sceKernelCheckThreadStackPlay Station Portable

Gets the remaining free size of the calling thread stack (?)

sceKernelGetThreadIdPlay Station Portable

Gets the thread UID of the calling thread.

sceKernelGetTlsAddrPlay Station Portable

Gets the address of a user TLS memory pool.

sceKernelLockLwMutexPlay Station Portable

Locks a lightweight mutex a number of times.

sceKernelLockLwMutexCBPlay Station Portable

Locks a lightweight mutex a number of times, but service any callbacks as necessary.

sceKernelMemcpy⚠Play Station Portable

What can I say? It is memcpy with a weird prefix.

sceKernelMemset⚠Play Station Portable

What can I say? It is memset with a weird prefix.

sceKernelReferLwMutexStatusPlay Station Portable

Gets the current state of a lightweight mutex.

sceKernelTryLockLwMutexPlay Station Portable

Tries to lock a lightweight mutex a number of times.

sceKernelUnlockLwMutexPlay Station Portable

Unlocks a mutex a number of times.